@@ -12,39 +12,39 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.10",
-    "@angular/common": "^7.2.10",
-    "@angular/compiler": "^7.2.10",
-    "@angular/core": "^7.2.10",
-    "@angular/platform-browser": "^7.2.10",
-    "@angular/platform-browser-dynamic": "^7.2.10",
-    "angular-calendar": "^0.27.0",
-    "core-js": "^2.5.6",
-    "date-fns": "^1.29.0",
-    "rxjs": "^6.1.0",
-    "tslib": "^1.9.0",
-    "zone.js": "^0.8.26"
+    "@angular/animations": "^8.2.14",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "angular-calendar": "^0.28.2",
+    "core-js": "^3.4.7",
+    "date-fns": "^2.8.1",
+    "rxjs": "^6.5.3",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "^7.3.6",
-    "@angular/compiler-cli": "^7.2.10",
-    "@angular/language-service": "^7.2.10",
-    "@types/jasmine": "^3.3.12",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "^11.11.5",
-    "codelyzer": "^4.4.4",
-    "jasmine-core": "^3.2.1",
+    "@angular-devkit/build-angular": "~0.803.20",
+    "@angular/cli": "^8.3.20",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
+    "@types/jasmine": "^3.5.0",
+    "@types/jasminewd2": "^2.0.8",
+    "@types/node": "^12.12.14",
+    "codelyzer": "^5.0.1",
+    "jasmine-core": "^3.5.0",
     "jasmine-spec-reporter": "^4.2.1",
-    "karma": "^4.0.1",
-    "karma-chrome-launcher": "^2.2.0",
+    "karma": "^4.4.1",
+    "karma-chrome-launcher": "^3.1.0",
     "karma-cli": "^2.0.0",
-    "karma-coverage-istanbul-reporter": "^2.0.3",
+    "karma-coverage-istanbul-reporter": "^2.1.1",
     "karma-jasmine": "^2.0.1",
-    "karma-jasmine-html-reporter": "^1.3.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
     "protractor": "^5.4.0",
-    "ts-node": "^8.0.3",
-    "tslint": "^5.11.0",
-    "typescript": "~3.2.4"
+    "ts-node": "^8.5.4",
+    "tslint": "^5.20.1",
+    "typescript": "~3.5.3"
   }
 }
