@@ -5,30 +5,30 @@
   "author": "Matt Lewis",
   "license": "MIT",
   "dependencies": {
-    "@angular/animations": "^7.2.10",
-    "@angular/common": "^7.2.10",
-    "@angular/compiler": "^7.2.10",
-    "@angular/core": "^7.2.10",
-    "@angular/forms": "^7.2.10",
-    "@angular/platform-browser": "^7.2.10",
-    "@angular/platform-browser-dynamic": "^7.2.10",
-    "angular-calendar": "^0.27.0",
-    "core-js": "^2.5.7",
-    "date-fns": "^1.29.0",
+    "@angular/animations": "^8.2.14",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "angular-calendar": "^0.28.2",
+    "core-js": "^2.6.10",
+    "date-fns": "^2.8.1",
     "reflect-metadata": "^0.1.12",
-    "rxjs": "^6.3.1",
-    "zone.js": "^0.8.26"
+    "rxjs": "^6.5.3",
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@types/core-js": "^2.5.0",
-    "@types/node": "^11.11.5",
-    "lite-server": "^2.4.0",
-    "rollup": "^1.7.0",
-    "rollup-plugin-commonjs": "^9.1.6",
+    "@types/core-js": "^2.5.2",
+    "@types/node": "^12.12.14",
+    "lite-server": "^2.5.4",
+    "rollup": "^1.27.8",
+    "rollup-plugin-commonjs": "^10.1.0",
     "rollup-plugin-css-only": "^1.0.0",
-    "rollup-plugin-node-resolve": "^4.0.1",
-    "rollup-plugin-typescript": "^0.8.1",
-    "typescript": "~3.2.4"
+    "rollup-plugin-node-resolve": "^5.2.0",
+    "rollup-plugin-typescript": "^1.0.1",
+    "typescript": "~3.5.3"
   },
   "scripts": {
     "rollup": "rollup -f iife -c rollup.config.js -o dist/bundle.js",
