@@ -5,27 +5,27 @@
   "author": "Matt Lewis",
   "license": "MIT",
   "dependencies": {
-    "@angular/animations": "^7.2.10",
-    "@angular/common": "^7.2.10",
-    "@angular/compiler": "^7.2.10",
-    "@angular/core": "^7.2.10",
-    "@angular/forms": "^7.2.10",
-    "@angular/platform-browser": "^7.2.10",
-    "@angular/platform-browser-dynamic": "^7.2.10",
-    "angular-calendar": "^0.27.0",
-    "angular-in-memory-web-api": "^0.8.0",
-    "core-js": "^2.5.7",
-    "date-fns": "^1.29.0",
+    "@angular/animations": "^8.2.14",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "angular-calendar": "^0.28.2",
+    "angular-in-memory-web-api": "^0.9.0",
+    "core-js": "^2.6.10",
+    "date-fns": "^2.8.1",
     "reflect-metadata": "^0.1.12",
-    "rxjs": "^6.3.1",
+    "rxjs": "^6.5.3",
     "systemjs": "^0.19.39",
-    "zone.js": "^0.8.26"
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@types/node": "^11.11.5",
-    "concurrently": "^4.0.1",
-    "lite-server": "^2.4.0",
-    "typescript": "~3.2.4"
+    "@types/node": "^12.12.14",
+    "concurrently": "^5.0.0",
+    "lite-server": "^2.5.4",
+    "typescript": "~3.5.3"
   },
   "scripts": {
     "start": "tsc && concurrently \"tsc -w\" \"lite-server\" ",
