@@ -12,15 +12,7 @@ import {
 } from '@angular/core';
 import { DOCUMENT } from '@angular/common';
 import { Observable, Subject } from 'rxjs';
-import { requestIdleCallbackObservable } from './request-idle-callback';
-import { switchMapTo, takeUntil } from 'rxjs/operators';
-
-const clickElements = new Set<HTMLElement>();
-
-const eventName: string =
-  typeof window !== 'undefined' && typeof window['Hammer'] !== 'undefined'
-    ? 'tap'
-    : 'click';
+import { takeUntil } from 'rxjs/operators';
 
 @Directive({
   selector: '[mwlClick]'
@@ -35,53 +27,27 @@ export class ClickDirective implements OnInit, OnDestroy {
   constructor(
     private renderer: Renderer2,
     private elm: ElementRef<HTMLElement>,
-    @Inject(DOCUMENT) private document,
-    private zone: NgZone
+    @Inject(DOCUMENT) private document
   ) {}
 
   ngOnInit(): void {
     if (!this.clickListenerDisabled) {
-      this.renderer.setAttribute(
-        this.elm.nativeElement,
-        'data-calendar-clickable',
-        'true'
-      );
-      clickElements.add(this.elm.nativeElement);
-
-      // issue #942 - lazily initialise all click handlers after initial render as hammerjs is slow
-      requestIdleCallbackObservable()
-        .pipe(
-          switchMapTo(this.listen()),
-          takeUntil(this.destroy$)
-        )
+      this.listen()
+        .pipe(takeUntil(this.destroy$))
         .subscribe(event => {
-          // prevent child click events from firing on parent elements that also have click events
-          let nearestClickableParent = event.target as HTMLElement;
-          while (
-            !clickElements.has(nearestClickableParent) &&
-            nearestClickableParent !== this.document.body
-          ) {
-            nearestClickableParent = nearestClickableParent.parentElement;
-          }
-          const isThisClickableElement =
-            this.elm.nativeElement === nearestClickableParent;
-          if (isThisClickableElement) {
-            this.zone.run(() => {
-              this.click.next(event);
-            });
-          }
+          event.stopPropagation();
+          this.click.emit(event);
         });
     }
   }
 
   ngOnDestroy(): void {
     this.destroy$.next();
-    clickElements.delete(this.elm.nativeElement);
   }
 
   private listen() {
     return new Observable<MouseEvent>(observer => {
-      return this.renderer.listen(this.elm.nativeElement, eventName, event => {
+      return this.renderer.listen(this.elm.nativeElement, 'click', event => {
         observer.next(event);
       });
     });
