@@ -5,63 +5,63 @@ export function adapterFactory(moment): DateAdapter {
   return {
     ...baseAdapterFactory(moment),
 
-    addWeeks(date: Date | string | number, amount: number): Date {
+    addWeeks(date: Date | number, amount: number): Date {
       return moment(date)
         .add(amount, 'weeks')
         .toDate();
     },
 
-    addMonths(date: Date | string | number, amount: number): Date {
+    addMonths(date: Date | number, amount: number): Date {
       return moment(date)
         .add(amount, 'months')
         .toDate();
     },
 
-    subDays(date: Date | string | number, amount: number): Date {
+    subDays(date: Date | number, amount: number): Date {
       return moment(date)
         .subtract(amount, 'days')
         .toDate();
     },
 
-    subWeeks(date: Date | string | number, amount: number): Date {
+    subWeeks(date: Date | number, amount: number): Date {
       return moment(date)
         .subtract(amount, 'weeks')
         .toDate();
     },
 
-    subMonths(date: Date | string | number, amount: number): Date {
+    subMonths(date: Date | number, amount: number): Date {
       return moment(date)
         .subtract(amount, 'months')
         .toDate();
     },
 
-    getISOWeek(date: Date | string | number): number {
+    getISOWeek(date: Date | number): number {
       return moment(date).isoWeek();
     },
 
-    setDate(date: Date | string | number, dayOfMonth: number): Date {
+    setDate(date: Date | number, dayOfMonth: number): Date {
       return moment(date)
         .date(dayOfMonth)
         .toDate();
     },
 
-    setMonth(date: Date | string | number, month: number): Date {
+    setMonth(date: Date | number, month: number): Date {
       return moment(date)
         .month(month)
         .toDate();
     },
 
-    setYear(date: Date | string | number, year: number): Date {
+    setYear(date: Date | number, year: number): Date {
       return moment(date)
         .year(year)
         .toDate();
     },
 
-    getDate(date: Date | string | number): number {
+    getDate(date: Date | number): number {
       return moment(date).date();
     },
 
-    getYear(date: Date | string | number): number {
+    getYear(date: Date | number): number {
       return moment(date).year();
     }
   };
