@@ -1,5 +1,5 @@
 import { ModuleWithProviders, NgModule, Provider } from '@angular/core';
-import { CommonModule } from '@angular/common';
+import { CommonModule, I18nPluralPipe } from '@angular/common';
 import { CalendarEventActionsComponent } from './calendar-event-actions.component';
 import { CalendarEventTitleComponent } from './calendar-event-title.component';
 import {
@@ -12,14 +12,18 @@ import { CalendarTodayDirective } from './calendar-today.directive';
 import { CalendarDatePipe } from './calendar-date.pipe';
 import { CalendarEventTitlePipe } from './calendar-event-title.pipe';
 import { ClickDirective } from './click.directive';
+import { KeydownEnterDirective } from './keydown-enter.directive';
 import { CalendarEventTitleFormatter } from './calendar-event-title-formatter.provider';
 import { CalendarDateFormatter } from './calendar-date-formatter.provider';
 import { CalendarUtils } from './calendar-utils.provider';
+import { CalendarA11y } from './calendar-a11y.provider';
+import { CalendarA11yPipe } from './calendar-a11y.pipe';
 
 export interface CalendarModuleConfig {
   eventTitleFormatter?: Provider;
   dateFormatter?: Provider;
   utils?: Provider;
+  a11y?: Provider;
 }
 
 export * from './calendar-event-title-formatter.provider';
@@ -28,6 +32,8 @@ export * from './calendar-native-date-formatter.provider';
 export * from './calendar-angular-date-formatter.provider';
 export * from './calendar-date-formatter.provider';
 export * from './calendar-utils.provider';
+export * from './calendar-a11y.provider';
+export * from './calendar-a11y.interface';
 export * from './calendar-date-formatter.interface';
 export * from './calendar-event-times-changed-event.interface';
 export * from '../../date-adapters/date-adapter';
@@ -67,7 +73,9 @@ export {
     CalendarTodayDirective,
     CalendarDatePipe,
     CalendarEventTitlePipe,
-    ClickDirective
+    CalendarA11yPipe,
+    ClickDirective,
+    KeydownEnterDirective
   ],
   imports: [CommonModule],
   exports: [
@@ -80,8 +88,11 @@ export {
     CalendarTodayDirective,
     CalendarDatePipe,
     CalendarEventTitlePipe,
-    ClickDirective
+    CalendarA11yPipe,
+    ClickDirective,
+    KeydownEnterDirective
   ],
+  providers: [I18nPluralPipe],
   entryComponents: [CalendarTooltipWindowComponent]
 })
 export class CalendarCommonModule {
@@ -95,7 +106,8 @@ export class CalendarCommonModule {
         dateAdapter,
         config.eventTitleFormatter || CalendarEventTitleFormatter,
         config.dateFormatter || CalendarDateFormatter,
-        config.utils || CalendarUtils
+        config.utils || CalendarUtils,
+        config.a11y || CalendarA11y
       ]
     };
   }
