@@ -1,8 +1,8 @@
 import {
   CalendarEvent,
-  DayViewEvent,
-  DayViewHour,
-  DayViewHourSegment,
+  WeekViewTimeEvent,
+  WeekViewHour,
+  WeekViewHourSegment,
   validateEvents as validateEventsWithoutLog,
   ViewPeriod,
   WeekDay,
@@ -40,18 +40,30 @@ export const trackByWeekDayHeaderDate = (index: number, day: WeekDay) =>
 
 export const trackByHourSegment = (
   index: number,
-  segment: DayViewHourSegment
+  segment: WeekViewHourSegment
 ) => segment.date.toISOString();
 
-export const trackByHour = (index: number, hour: DayViewHour) =>
+export const trackByHour = (index: number, hour: WeekViewHour) =>
   hour.segments[0].date.toISOString();
 
-export const trackByDayOrWeekEvent = (
+export const trackByWeekAllDayEvent = (
   index: number,
-  weekEvent: WeekViewAllDayEvent | DayViewEvent
+  weekEvent: WeekViewAllDayEvent
+) => (weekEvent.event.id ? weekEvent.event.id : weekEvent.event);
+
+export const trackByWeekTimeEvent = (
+  index: number,
+  weekEvent: WeekViewTimeEvent
 ) => (weekEvent.event.id ? weekEvent.event.id : weekEvent.event);
 
 const MINUTES_IN_HOUR = 60;
+
+function getPixelAmountInMinutes(
+  hourSegments: number,
+  hourSegmentHeight: number
+) {
+  return MINUTES_IN_HOUR / (hourSegments * hourSegmentHeight);
+}
 
 export function getMinutesMoved(
   movedY: number,
@@ -63,8 +75,10 @@ export function getMinutesMoved(
     movedY,
     eventSnapSize || hourSegmentHeight
   );
-  const pixelAmountInMinutes =
-    MINUTES_IN_HOUR / (hourSegments * hourSegmentHeight);
+  const pixelAmountInMinutes = getPixelAmountInMinutes(
+    hourSegments,
+    hourSegmentHeight
+  );
   return draggedInPixelsSnapSize * pixelAmountInMinutes;
 }
 
@@ -73,7 +87,7 @@ export function getMinimumEventHeightInMinutes(
   hourSegmentHeight: number
 ) {
   return (
-    (MINUTES_IN_HOUR / (hourSegments * hourSegmentHeight)) * hourSegmentHeight
+    getPixelAmountInMinutes(hourSegments, hourSegmentHeight) * hourSegmentHeight
   );
 }
 
