@@ -1,11 +1,8 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { ResizableModule } from 'angular-resizable-element';
-import { DragAndDropModule } from 'angular-draggable-droppable';
 import { CalendarDayViewComponent } from './calendar-day-view.component';
-import { CalendarDayViewHourSegmentComponent } from './calendar-day-view-hour-segment.component';
-import { CalendarDayViewEventComponent } from './calendar-day-view-event.component';
 import { CalendarCommonModule } from '../common/calendar-common.module';
+import { CalendarWeekModule } from '../week/calendar-week.module';
 
 export {
   CalendarDayViewComponent,
@@ -13,23 +10,8 @@ export {
 } from './calendar-day-view.component';
 
 @NgModule({
-  imports: [
-    CommonModule,
-    ResizableModule,
-    DragAndDropModule,
-    CalendarCommonModule
-  ],
-  declarations: [
-    CalendarDayViewComponent,
-    CalendarDayViewHourSegmentComponent,
-    CalendarDayViewEventComponent
-  ],
-  exports: [
-    ResizableModule,
-    DragAndDropModule,
-    CalendarDayViewComponent,
-    CalendarDayViewHourSegmentComponent,
-    CalendarDayViewEventComponent
-  ]
+  imports: [CommonModule, CalendarCommonModule, CalendarWeekModule],
+  declarations: [CalendarDayViewComponent],
+  exports: [CalendarDayViewComponent]
 })
 export class CalendarDayModule {}
