@@ -50,7 +50,27 @@ export const collapseAnimation: AnimationTriggerMetadata = trigger('collapse', [
       let-trackByEventId="trackByEventId"
       let-validateDrag="validateDrag"
     >
-      <div class="cal-open-day-events" [@collapse] *ngIf="isOpen">
+      <div
+        class="cal-open-day-events"
+        [@collapse]
+        *ngIf="isOpen"
+        role="application"
+      >
+        <span
+          tabindex="-1"
+          role="alert"
+          [attr.aria-label]="
+            { date: date, locale: locale } | calendarA11y: 'openDayEventsAlert'
+          "
+        ></span>
+        <span
+          tabindex="0"
+          role="landmark"
+          [attr.aria-label]="
+            { date: date, locale: locale }
+              | calendarA11y: 'openDayEventsLandmark'
+          "
+        ></span>
         <div
           *ngFor="let event of events; trackBy: trackByEventId"
           [ngClass]="event?.cssClass"
@@ -71,7 +91,17 @@ export const collapseAnimation: AnimationTriggerMetadata = trigger('collapse', [
             [event]="event"
             [customTemplate]="eventTitleTemplate"
             view="month"
-            (mwlClick)="eventClicked.emit({ event: event })"
+            (mwlClick)="
+              eventClicked.emit({ event: event, sourceEvent: $event })
+            "
+            (mwlKeydownEnter)="
+              eventClicked.emit({ event: event, sourceEvent: $event })
+            "
+            tabindex="0"
+            [attr.aria-label]="
+              { event: event, locale: locale }
+                | calendarA11y: 'eventDescription'
+            "
           >
           </mwl-calendar-event-title>
           &ngsp;
@@ -98,6 +128,8 @@ export const collapseAnimation: AnimationTriggerMetadata = trigger('collapse', [
   animations: [collapseAnimation]
 })
 export class CalendarOpenDayEventsComponent {
+  @Input() locale: string;
+
   @Input() isOpen: boolean = false;
 
   @Input() events: CalendarEvent[];
@@ -108,9 +140,12 @@ export class CalendarOpenDayEventsComponent {
 
   @Input() eventActionsTemplate: TemplateRef<any>;
 
+  @Input() date: Date;
+
   @Output()
-  eventClicked: EventEmitter<{ event: CalendarEvent }> = new EventEmitter<{
+  eventClicked = new EventEmitter<{
     event: CalendarEvent;
+    sourceEvent: MouseEvent | KeyboardEvent;
   }>();
 
   trackByEventId = trackByEventId;
