@@ -7,7 +7,7 @@ import {
 } from '@angular/core';
 import {
   WeekViewAllDayEvent,
-  DayViewEvent,
+  WeekViewTimeEvent,
   WeekViewHourColumn
 } from 'calendar-utils';
 import { PlacementArray } from 'positioning';
@@ -25,6 +25,7 @@ import { PlacementArray } from 'positioning';
       let-tooltipDisabled="tooltipDisabled"
       let-tooltipDelay="tooltipDelay"
       let-column="column"
+      let-daysInWeek="daysInWeek"
     >
       <div
         class="cal-event"
@@ -35,7 +36,9 @@ import { PlacementArray } from 'positioning';
         [mwlCalendarTooltip]="
           !tooltipDisabled
             ? (weekEvent.event.title
-              | calendarEventTitle: 'weekTooltip':weekEvent.event)
+              | calendarEventTitle
+                : (daysInWeek === 1 ? 'dayTooltip' : 'weekTooltip')
+                : weekEvent.event)
             : ''
         "
         [tooltipPlacement]="tooltipPlacement"
@@ -43,7 +46,14 @@ import { PlacementArray } from 'positioning';
         [tooltipTemplate]="tooltipTemplate"
         [tooltipAppendToBody]="tooltipAppendToBody"
         [tooltipDelay]="tooltipDelay"
-        (mwlClick)="eventClicked.emit()"
+        (mwlClick)="eventClicked.emit({ sourceEvent: $event })"
+        (mwlKeydownEnter)="eventClicked.emit({ sourceEvent: $event })"
+        tabindex="0"
+        role="application"
+        [attr.aria-label]="
+          { event: weekEvent.event, locale: locale }
+            | calendarA11y: 'eventDescription'
+        "
       >
         <mwl-calendar-event-actions
           [event]="weekEvent.event"
@@ -54,7 +64,7 @@ import { PlacementArray } from 'positioning';
         <mwl-calendar-event-title
           [event]="weekEvent.event"
           [customTemplate]="eventTitleTemplate"
-          view="week"
+          [view]="daysInWeek === 1 ? 'day' : 'week'"
         >
         </mwl-calendar-event-title>
       </div>
@@ -69,14 +79,17 @@ import { PlacementArray } from 'positioning';
         tooltipAppendToBody: tooltipAppendToBody,
         tooltipDisabled: tooltipDisabled,
         tooltipDelay: tooltipDelay,
-        column: column
+        column: column,
+        daysInWeek: daysInWeek
       }"
     >
     </ng-template>
   `
 })
 export class CalendarWeekViewEventComponent {
-  @Input() weekEvent: WeekViewAllDayEvent | DayViewEvent;
+  @Input() locale: string;
+
+  @Input() weekEvent: WeekViewAllDayEvent | WeekViewTimeEvent;
 
   @Input() tooltipPlacement: PlacementArray;
 
@@ -96,5 +109,9 @@ export class CalendarWeekViewEventComponent {
 
   @Input() column: WeekViewHourColumn;
 
-  @Output() eventClicked: EventEmitter<any> = new EventEmitter();
+  @Input() daysInWeek: number;
+
+  @Output() eventClicked = new EventEmitter<{
+    sourceEvent: MouseEvent | KeyboardEvent;
+  }>();
 }
