@@ -19,9 +19,9 @@ import {
   WeekView,
   ViewPeriod,
   WeekViewHourColumn,
-  DayViewEvent,
-  DayViewHourSegment,
-  DayViewHour,
+  WeekViewTimeEvent,
+  WeekViewHourSegment,
+  WeekViewHour,
   WeekViewAllDayEventRow
 } from 'calendar-utils';
 import { ResizeEvent } from 'angular-resizable-element';
@@ -42,10 +42,11 @@ import {
   getDefaultEventEnd,
   getMinimumEventHeightInMinutes,
   addDaysWithExclusions,
-  trackByDayOrWeekEvent,
   isDraggedWithinPeriod,
   shouldFireDroppedEvent,
-  getWeekViewPeriod
+  getWeekViewPeriod,
+  trackByWeekAllDayEvent,
+  trackByWeekTimeEvent
 } from '../common/util';
 import { DateAdapter } from '../../date-adapters/date-adapter';
 import {
@@ -79,7 +80,7 @@ export interface CalendarWeekViewBeforeRenderEvent extends WeekView {
 @Component({
   selector: 'mwl-calendar-week-view',
   template: `
-    <div class="cal-week-view">
+    <div class="cal-week-view" role="grid">
       <mwl-calendar-week-view-header
         [days]="days"
         [locale]="locale"
@@ -88,6 +89,7 @@ export interface CalendarWeekViewBeforeRenderEvent extends WeekView {
         (eventDropped)="
           eventDropped({ dropData: $event }, $event.newStart, true)
         "
+        (dragEnter)="dateDragEnter($event.date)"
       >
       </mwl-calendar-week-view-header>
       <div
@@ -109,6 +111,7 @@ export interface CalendarWeekViewBeforeRenderEvent extends WeekView {
             mwlDroppable
             dragOverClass="cal-drag-over"
             (drop)="eventDropped($event, day.date, true)"
+            (dragEnter)="dateDragEnter(day.date)"
           ></div>
         </div>
         <div
@@ -119,7 +122,7 @@ export interface CalendarWeekViewBeforeRenderEvent extends WeekView {
           <div
             *ngFor="
               let allDayEvent of eventRow.row;
-              trackBy: trackByDayOrWeekEvent
+              trackBy: trackByWeekAllDayEvent
             "
             #event
             class="cal-event-container"
@@ -167,6 +170,7 @@ export interface CalendarWeekViewBeforeRenderEvent extends WeekView {
               [resizeEdges]="{ left: true }"
             ></div>
             <mwl-calendar-week-view-event
+              [locale]="locale"
               [weekEvent]="allDayEvent"
               [tooltipPlacement]="tooltipPlacement"
               [tooltipTemplate]="tooltipTemplate"
@@ -175,7 +179,13 @@ export interface CalendarWeekViewBeforeRenderEvent extends WeekView {
               [customTemplate]="eventTemplate"
               [eventTitleTemplate]="eventTitleTemplate"
               [eventActionsTemplate]="eventActionsTemplate"
-              (eventClicked)="eventClicked.emit({ event: allDayEvent.event })"
+              [daysInWeek]="daysInWeek"
+              (eventClicked)="
+                eventClicked.emit({
+                  event: allDayEvent.event,
+                  sourceEvent: $event.sourceEvent
+                })
+              "
             >
             </mwl-calendar-week-view-event>
             <div
@@ -196,7 +206,10 @@ export interface CalendarWeekViewBeforeRenderEvent extends WeekView {
         (dragEnter)="dragEnter('time')"
         (dragLeave)="dragLeave('time')"
       >
-        <div class="cal-time-label-column" *ngIf="view.hourColumns.length > 0">
+        <div
+          class="cal-time-label-column"
+          *ngIf="view.hourColumns.length > 0 && daysInWeek !== 1"
+        >
           <div
             *ngFor="
               let hour of view.hourColumns[0].hours;
@@ -214,6 +227,7 @@ export interface CalendarWeekViewBeforeRenderEvent extends WeekView {
               [locale]="locale"
               [customTemplate]="hourSegmentTemplate"
               [isTimeLabel]="true"
+              [daysInWeek]="daysInWeek"
             >
             </mwl-calendar-week-view-hour-segment>
           </div>
@@ -227,94 +241,116 @@ export interface CalendarWeekViewBeforeRenderEvent extends WeekView {
             class="cal-day-column"
             *ngFor="let column of view.hourColumns; trackBy: trackByHourColumn"
           >
-            <div
-              *ngFor="
-                let timeEvent of column.events;
-                trackBy: trackByDayOrWeekEvent
-              "
-              #event
-              class="cal-event-container"
-              [class.cal-draggable]="
-                timeEvent.event.draggable && timeEventResizes.size === 0
-              "
-              [class.cal-starts-within-day]="!timeEvent.startsBeforeDay"
-              [class.cal-ends-within-day]="!timeEvent.endsAfterDay"
-              [ngClass]="timeEvent.event.cssClass"
-              [hidden]="timeEvent.height === 0 && timeEvent.width === 0"
-              [style.top.px]="timeEvent.top"
-              [style.height.px]="timeEvent.height"
-              [style.left.%]="timeEvent.left"
-              [style.width.%]="timeEvent.width"
-              mwlResizable
-              [resizeSnapGrid]="{
-                left: dayColumnWidth,
-                right: dayColumnWidth,
-                top: eventSnapSize || hourSegmentHeight,
-                bottom: eventSnapSize || hourSegmentHeight
-              }"
-              [validateResize]="validateResize"
-              [allowNegativeResizes]="true"
-              (resizeStart)="
-                timeEventResizeStarted(dayColumns, timeEvent, $event)
-              "
-              (resizing)="timeEventResizing(timeEvent, $event)"
-              (resizeEnd)="timeEventResizeEnded(timeEvent)"
-              mwlDraggable
-              dragActiveClass="cal-drag-active"
-              [dropData]="{ event: timeEvent.event, calendarId: calendarId }"
-              [dragAxis]="{
-                x: timeEvent.event.draggable && timeEventResizes.size === 0,
-                y: timeEvent.event.draggable && timeEventResizes.size === 0
-              }"
-              [dragSnapGrid]="
-                snapDraggedEvents
-                  ? { x: dayColumnWidth, y: eventSnapSize || hourSegmentHeight }
-                  : {}
-              "
-              [ghostDragEnabled]="!snapDraggedEvents"
-              [validateDrag]="validateDrag"
-              (dragStart)="dragStarted(dayColumns, event, timeEvent)"
-              (dragging)="dragMove(timeEvent, $event)"
-              (dragEnd)="dragEnded(timeEvent, $event, dayColumnWidth, true)"
-            >
+            <mwl-calendar-week-view-current-time-marker
+              [columnDate]="column.date"
+              [dayStartHour]="dayStartHour"
+              [dayStartMinute]="dayStartMinute"
+              [dayEndHour]="dayEndHour"
+              [dayEndMinute]="dayEndMinute"
+              [hourSegments]="hourSegments"
+              [hourSegmentHeight]="hourSegmentHeight"
+              [customTemplate]="currentTimeMarkerTemplate"
+            ></mwl-calendar-week-view-current-time-marker>
+            <div class="cal-events-container">
               <div
-                class="cal-resize-handle cal-resize-handle-before-start"
-                *ngIf="
-                  timeEvent.event?.resizable?.beforeStart &&
-                  !timeEvent.startsBeforeDay
+                *ngFor="
+                  let timeEvent of column.events;
+                  trackBy: trackByWeekTimeEvent
                 "
-                mwlResizeHandle
-                [resizeEdges]="{
-                  left: true,
-                  top: true
+                #event
+                class="cal-event-container"
+                [class.cal-draggable]="
+                  timeEvent.event.draggable && timeEventResizes.size === 0
+                "
+                [class.cal-starts-within-day]="!timeEvent.startsBeforeDay"
+                [class.cal-ends-within-day]="!timeEvent.endsAfterDay"
+                [ngClass]="timeEvent.event.cssClass"
+                [hidden]="timeEvent.height === 0 && timeEvent.width === 0"
+                [style.top.px]="timeEvent.top"
+                [style.height.px]="timeEvent.height"
+                [style.left.%]="timeEvent.left"
+                [style.width.%]="timeEvent.width"
+                mwlResizable
+                [resizeSnapGrid]="{
+                  left: dayColumnWidth,
+                  right: dayColumnWidth,
+                  top: eventSnapSize || hourSegmentHeight,
+                  bottom: eventSnapSize || hourSegmentHeight
                 }"
-              ></div>
-              <mwl-calendar-week-view-event
-                [weekEvent]="timeEvent"
-                [tooltipPlacement]="tooltipPlacement"
-                [tooltipTemplate]="tooltipTemplate"
-                [tooltipAppendToBody]="tooltipAppendToBody"
-                [tooltipDisabled]="dragActive || timeEventResizes.size > 0"
-                [tooltipDelay]="tooltipDelay"
-                [customTemplate]="eventTemplate"
-                [eventTitleTemplate]="eventTitleTemplate"
-                [eventActionsTemplate]="eventActionsTemplate"
-                [column]="column"
-                (eventClicked)="eventClicked.emit({ event: timeEvent.event })"
+                [validateResize]="validateResize"
+                [allowNegativeResizes]="true"
+                (resizeStart)="
+                  timeEventResizeStarted(dayColumns, timeEvent, $event)
+                "
+                (resizing)="timeEventResizing(timeEvent, $event)"
+                (resizeEnd)="timeEventResizeEnded(timeEvent)"
+                mwlDraggable
+                dragActiveClass="cal-drag-active"
+                [dropData]="{ event: timeEvent.event, calendarId: calendarId }"
+                [dragAxis]="{
+                  x: timeEvent.event.draggable && timeEventResizes.size === 0,
+                  y: timeEvent.event.draggable && timeEventResizes.size === 0
+                }"
+                [dragSnapGrid]="
+                  snapDraggedEvents
+                    ? {
+                        x: dayColumnWidth,
+                        y: eventSnapSize || hourSegmentHeight
+                      }
+                    : {}
+                "
+                [ghostDragEnabled]="!snapDraggedEvents"
+                [validateDrag]="validateDrag"
+                (dragStart)="dragStarted(dayColumns, event, timeEvent)"
+                (dragging)="dragMove(timeEvent, $event)"
+                (dragEnd)="dragEnded(timeEvent, $event, dayColumnWidth, true)"
               >
-              </mwl-calendar-week-view-event>
-              <div
-                class="cal-resize-handle cal-resize-handle-after-end"
-                *ngIf="
-                  timeEvent.event?.resizable?.afterEnd &&
-                  !timeEvent.endsAfterDay
-                "
-                mwlResizeHandle
-                [resizeEdges]="{
-                  right: true,
-                  bottom: true
-                }"
-              ></div>
+                <div
+                  class="cal-resize-handle cal-resize-handle-before-start"
+                  *ngIf="
+                    timeEvent.event?.resizable?.beforeStart &&
+                    !timeEvent.startsBeforeDay
+                  "
+                  mwlResizeHandle
+                  [resizeEdges]="{
+                    left: true,
+                    top: true
+                  }"
+                ></div>
+                <mwl-calendar-week-view-event
+                  [locale]="locale"
+                  [weekEvent]="timeEvent"
+                  [tooltipPlacement]="tooltipPlacement"
+                  [tooltipTemplate]="tooltipTemplate"
+                  [tooltipAppendToBody]="tooltipAppendToBody"
+                  [tooltipDisabled]="dragActive || timeEventResizes.size > 0"
+                  [tooltipDelay]="tooltipDelay"
+                  [customTemplate]="eventTemplate"
+                  [eventTitleTemplate]="eventTitleTemplate"
+                  [eventActionsTemplate]="eventActionsTemplate"
+                  [column]="column"
+                  [daysInWeek]="daysInWeek"
+                  (eventClicked)="
+                    eventClicked.emit({
+                      event: timeEvent.event,
+                      sourceEvent: $event.sourceEvent
+                    })
+                  "
+                >
+                </mwl-calendar-week-view-event>
+                <div
+                  class="cal-resize-handle cal-resize-handle-after-end"
+                  *ngIf="
+                    timeEvent.event?.resizable?.afterEnd &&
+                    !timeEvent.endsAfterDay
+                  "
+                  mwlResizeHandle
+                  [resizeEdges]="{
+                    right: true,
+                    bottom: true
+                  }"
+                ></div>
+              </div>
             </div>
 
             <div
@@ -336,7 +372,13 @@ export interface CalendarWeekViewBeforeRenderEvent extends WeekView {
                 [segmentHeight]="hourSegmentHeight"
                 [locale]="locale"
                 [customTemplate]="hourSegmentTemplate"
-                (mwlClick)="hourSegmentClicked.emit({ date: segment.date })"
+                [daysInWeek]="daysInWeek"
+                (mwlClick)="
+                  hourSegmentClicked.emit({
+                    date: segment.date,
+                    sourceEvent: $event
+                  })
+                "
                 [clickListenerDisabled]="
                   hourSegmentClicked.observers.length === 0
                 "
@@ -346,6 +388,8 @@ export interface CalendarWeekViewBeforeRenderEvent extends WeekView {
                 "
                 dragActiveClass="cal-drag-active"
                 (drop)="eventDropped($event, segment.date, false)"
+                (dragEnter)="dateDragEnter(segment.date)"
+                [isTimeLabel]="daysInWeek === 1"
               >
               </mwl-calendar-week-view-hour-segment>
             </div>
@@ -498,11 +542,17 @@ export class CalendarWeekViewComponent implements OnChanges, OnInit, OnDestroy {
   @Input() daysInWeek: number;
 
   /**
+   * A custom template to use for the current time marker
+   */
+  @Input() currentTimeMarkerTemplate: TemplateRef<any>;
+
+  /**
    * Called when a header week day is clicked. Adding a `cssClass` property on `$event.day` will add that class to the header element
    */
   @Output()
   dayHeaderClicked = new EventEmitter<{
     day: WeekDay;
+    sourceEvent: MouseEvent;
   }>();
 
   /**
@@ -511,6 +561,7 @@ export class CalendarWeekViewComponent implements OnChanges, OnInit, OnDestroy {
   @Output()
   eventClicked = new EventEmitter<{
     event: CalendarEvent;
+    sourceEvent: MouseEvent | KeyboardEvent;
   }>();
 
   /**
@@ -532,6 +583,7 @@ export class CalendarWeekViewComponent implements OnChanges, OnInit, OnDestroy {
   @Output()
   hourSegmentClicked = new EventEmitter<{
     date: Date;
+    sourceEvent: MouseEvent;
   }>();
 
   /**
@@ -618,7 +670,17 @@ export class CalendarWeekViewComponent implements OnChanges, OnInit, OnDestroy {
   /**
    * @hidden
    */
-  trackByDayOrWeekEvent = trackByDayOrWeekEvent;
+  trackByWeekAllDayEvent = trackByWeekAllDayEvent;
+
+  /**
+   * @hidden
+   */
+  trackByWeekTimeEvent = trackByWeekTimeEvent;
+
+  /**
+   * @hidden
+   */
+  private lastDragEnterDate: Date;
 
   /**
    * @hidden
@@ -722,7 +784,7 @@ export class CalendarWeekViewComponent implements OnChanges, OnInit, OnDestroy {
    */
   timeEventResizeStarted(
     eventsContainer: HTMLElement,
-    timeEvent: DayViewEvent,
+    timeEvent: WeekViewTimeEvent,
     resizeEvent: ResizeEvent
   ): void {
     this.timeEventResizes.set(timeEvent.event, resizeEvent);
@@ -732,7 +794,7 @@ export class CalendarWeekViewComponent implements OnChanges, OnInit, OnDestroy {
   /**
    * @hidden
    */
-  timeEventResizing(timeEvent: DayViewEvent, resizeEvent: ResizeEvent) {
+  timeEventResizing(timeEvent: WeekViewTimeEvent, resizeEvent: ResizeEvent) {
     this.timeEventResizes.set(timeEvent.event, resizeEvent);
     const adjustedEvents = new Map<CalendarEvent, CalendarEvent>();
 
@@ -755,7 +817,7 @@ export class CalendarWeekViewComponent implements OnChanges, OnInit, OnDestroy {
   /**
    * @hidden
    */
-  timeEventResizeEnded(timeEvent: DayViewEvent) {
+  timeEventResizeEnded(timeEvent: WeekViewTimeEvent) {
     this.view = this.getWeekView(this.events);
     const lastResizeEvent = this.timeEventResizes.get(timeEvent.event);
     if (lastResizeEvent) {
@@ -872,12 +934,22 @@ export class CalendarWeekViewComponent implements OnChanges, OnInit, OnDestroy {
   /**
    * @hidden
    */
+  dateDragEnter(date: Date) {
+    this.lastDragEnterDate = date;
+  }
+
+  /**
+   * @hidden
+   */
   eventDropped(
     dropEvent: DropEvent<{ event?: CalendarEvent; calendarId?: symbol }>,
     date: Date,
     allDay: boolean
   ): void {
-    if (shouldFireDroppedEvent(dropEvent, date, allDay, this.calendarId)) {
+    if (
+      shouldFireDroppedEvent(dropEvent, date, allDay, this.calendarId) &&
+      this.lastDragEnterDate.getTime() === date.getTime()
+    ) {
       this.eventTimesChanged.emit({
         type: CalendarEventTimesChangedEventType.Drop,
         event: dropEvent.dropData.event,
@@ -907,7 +979,7 @@ export class CalendarWeekViewComponent implements OnChanges, OnInit, OnDestroy {
   dragStarted(
     eventsContainer: HTMLElement,
     event: HTMLElement,
-    dayEvent?: DayViewEvent
+    dayEvent?: WeekViewTimeEvent
   ): void {
     this.dayColumnWidth = this.getDayColumnWidth(eventsContainer);
     const dragHelper: CalendarDragHelper = new CalendarDragHelper(
@@ -949,7 +1021,7 @@ export class CalendarWeekViewComponent implements OnChanges, OnInit, OnDestroy {
   /**
    * @hidden
    */
-  dragMove(dayEvent: DayViewEvent, dragEvent: DragMoveEvent) {
+  dragMove(dayEvent: WeekViewTimeEvent, dragEvent: DragMoveEvent) {
     if (this.snapDraggedEvents) {
       const newEventTimes = this.getDragMovedEventTimes(
         dayEvent,
@@ -984,7 +1056,7 @@ export class CalendarWeekViewComponent implements OnChanges, OnInit, OnDestroy {
    * @hidden
    */
   dragEnded(
-    weekEvent: WeekViewAllDayEvent | DayViewEvent,
+    weekEvent: WeekViewAllDayEvent | WeekViewTimeEvent,
     dragEndEvent: DragEndEvent,
     dayWidth: number,
     useY = false
@@ -1076,7 +1148,7 @@ export class CalendarWeekViewComponent implements OnChanges, OnInit, OnDestroy {
   }
 
   protected getDragMovedEventTimes(
-    weekEvent: WeekViewAllDayEvent | DayViewEvent,
+    weekEvent: WeekViewAllDayEvent | WeekViewTimeEvent,
     dragEndEvent: DragEndEvent | DragMoveEvent,
     dayWidth: number,
     useY: boolean
