@@ -54,10 +54,10 @@ describe('CalendarDayViewComponent component', () => {
     eventTitle = _eventTitle_;
   }));
 
-  it('should generate the day view', () => {
-    const fixture: ComponentFixture<
+  it('should generate the week view with one day visible', () => {
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-29');
     fixture.componentInstance.events = [
       {
@@ -69,19 +69,17 @@ describe('CalendarDayViewComponent component', () => {
         }
       }
     ];
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
-    expect(fixture.componentInstance.view.events.length).to.equal(1);
-    expect(fixture.componentInstance.view.events[0].event).to.equal(
-      fixture.componentInstance.events[0]
-    );
-    expect(fixture.componentInstance.hours.length).to.equal(24);
+    fixture.detectChanges();
+    expect(
+      fixture.nativeElement.querySelectorAll('.cal-day-column').length
+    ).to.equal(1);
   });
 
-  it('should generate the week view with default colors for events', () => {
-    const fixture: ComponentFixture<
+  it('should generate the day view with default colors for events', () => {
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
-    fixture.componentInstance.ngOnInit();
+    );
+
     fixture.componentInstance.viewDate = new Date('2016-06-01');
     fixture.componentInstance.events = [
       {
@@ -90,7 +88,7 @@ describe('CalendarDayViewComponent component', () => {
         title: 'foo'
       }
     ];
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
+
     fixture.detectChanges();
 
     const computedStyles: CSSStyleDeclaration = window.getComputedStyle(
@@ -108,10 +106,10 @@ describe('CalendarDayViewComponent component', () => {
     fixture.destroy();
   });
 
-  it('should call the event clicked callback', async(() => {
-    const fixture: ComponentFixture<
+  it('should call the event clicked callback', () => {
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-29');
     fixture.componentInstance.events = [
       {
@@ -123,20 +121,20 @@ describe('CalendarDayViewComponent component', () => {
         }
       }
     ];
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
-    fixture.detectChanges();
-    fixture.componentInstance.eventClicked.subscribe(val => {
-      expect(val).to.deep.equal({
-        event: fixture.componentInstance.events[0]
-      });
-    });
-    fixture.nativeElement.querySelector('.cal-event-title').click();
-  }));
 
-  it('should call the event clicked callback on all day events', async(() => {
-    const fixture: ComponentFixture<
+    fixture.detectChanges();
+    const spy = sinon.spy();
+    fixture.componentInstance.eventClicked.subscribe(spy);
+    fixture.nativeElement.querySelector('.cal-event-title').click();
+    const call = spy.getCall(0).args[0];
+    expect(call.event).to.equal(fixture.componentInstance.events[0]);
+    expect(call.sourceEvent).to.be.an.instanceOf(MouseEvent);
+  });
+
+  it('should call the event clicked callback on all day events', () => {
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-29');
     fixture.componentInstance.events = [
       {
@@ -149,20 +147,20 @@ describe('CalendarDayViewComponent component', () => {
         }
       }
     ];
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
+
     fixture.detectChanges();
-    fixture.componentInstance.eventClicked.subscribe(val => {
-      expect(val).to.deep.equal({
-        event: fixture.componentInstance.events[0]
-      });
-    });
+    const spy = sinon.spy();
+    fixture.componentInstance.eventClicked.subscribe(spy);
     fixture.nativeElement.querySelector('.cal-event-title').click();
-  }));
+    const call = spy.getCall(0).args[0];
+    expect(call.event).to.equal(fixture.componentInstance.events[0]);
+    expect(call.sourceEvent).to.be.an.instanceOf(MouseEvent);
+  });
 
   it('should add a custom CSS class to events', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-01');
     fixture.componentInstance.events = [
       {
@@ -175,7 +173,7 @@ describe('CalendarDayViewComponent component', () => {
         }
       }
     ];
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
+
     fixture.detectChanges();
     expect(
       fixture.nativeElement
@@ -186,9 +184,9 @@ describe('CalendarDayViewComponent component', () => {
   });
 
   it('should add a custom CSS class to all day events', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-01');
     fixture.componentInstance.events = [
       {
@@ -202,62 +200,41 @@ describe('CalendarDayViewComponent component', () => {
         allDay: true
       }
     ];
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
+
     fixture.detectChanges();
     expect(
       fixture.nativeElement
-        .querySelector('.cal-all-day-events mwl-calendar-day-view-event')
+        .querySelector('.cal-all-day-events .cal-event-container')
         .classList.contains('foo')
     ).to.equal(true);
     fixture.destroy();
   });
 
   it('should call the hour segment clicked callback', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-01');
-    fixture.componentInstance.ngOnChanges({ viewDate: {} });
-    fixture.detectChanges();
-    fixture.componentInstance.hourSegmentClicked.subscribe(val => {
-      expect(val).to.deep.equal({
-        date: moment('2016-06-01')
-          .startOf('day')
-          .add(1, 'hour')
-          .add(30, 'minutes')
-          .toDate()
-      });
-    });
-    fixture.nativeElement.querySelectorAll('.cal-hour-segment')[3].click();
-  });
 
-  it('should refresh the view when the refresh observable is emitted on', () => {
-    const fixture: ComponentFixture<
-      CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
-    fixture.componentInstance.refresh = new Subject();
-    fixture.componentInstance.ngOnInit();
-    fixture.componentInstance.viewDate = new Date('2016-06-01');
-    fixture.componentInstance.ngOnChanges({ viewDate: {} });
-    const event: CalendarEvent = {
-      start: new Date('2016-06-01'),
-      end: new Date('2016-06-02'),
-      title: 'foo',
-      color: {
-        primary: 'blue',
-        secondary: 'lightblue'
-      }
-    };
-    fixture.componentInstance.events.push(event);
-    fixture.componentInstance.refresh.next(true);
-    expect(fixture.componentInstance.view.events[0].event).to.deep.equal(event);
-    fixture.destroy();
+    fixture.detectChanges();
+    const spy = sinon.spy();
+    fixture.componentInstance.hourSegmentClicked.subscribe(spy);
+    fixture.nativeElement.querySelectorAll('.cal-hour-segment')[3].click();
+    const args = spy.getCall(0).args[0];
+    expect(args.date).to.deep.equal(
+      moment('2016-06-01')
+        .startOf('day')
+        .add(1, 'hour')
+        .add(30, 'minutes')
+        .toDate()
+    );
+    expect(args.sourceEvent instanceof MouseEvent).to.be.true;
   });
 
   it('should allow the event title to be customised by the calendarConfig provider', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     eventTitle.day = (event: CalendarEvent) => {
       return `foo ${event.title}`;
     };
@@ -273,7 +250,7 @@ describe('CalendarDayViewComponent component', () => {
         }
       }
     ];
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
+
     fixture.detectChanges();
     const title: HTMLElement = fixture.nativeElement.querySelector(
       '.cal-event-title'
@@ -281,35 +258,10 @@ describe('CalendarDayViewComponent component', () => {
     expect(title.innerHTML).to.equal('foo bar');
   });
 
-  it('should update the hour grid and event list when the day start changes', () => {
-    const fixture: ComponentFixture<
-      CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
-    fixture.componentInstance.viewDate = new Date('2016-06-29');
-    fixture.componentInstance.events = [
-      {
-        start: new Date('2016-06-29'),
-        title: 'foo',
-        color: {
-          primary: '',
-          secondary: ''
-        }
-      }
-    ];
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
-    fixture.detectChanges();
-    expect(fixture.componentInstance.hours.length).to.equal(24);
-    expect(fixture.componentInstance.view.events.length).to.equal(1);
-    fixture.componentInstance.dayStartHour = 6;
-    fixture.componentInstance.ngOnChanges({ dayStartHour: {} });
-    expect(fixture.componentInstance.hours.length).to.equal(18);
-    expect(fixture.componentInstance.view.events.length).to.equal(0);
-  });
-
   it('should add event actions to each event', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     const eventClicked = sinon.spy();
     fixture.componentInstance.eventClicked.subscribe(eventClicked);
@@ -331,7 +283,7 @@ describe('CalendarDayViewComponent component', () => {
         ]
       }
     ];
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
+
     fixture.detectChanges();
     const action: HTMLElement = fixture.nativeElement.querySelector(
       '.cal-event .cal-event-action'
@@ -339,55 +291,33 @@ describe('CalendarDayViewComponent component', () => {
     expect(action.innerHTML).to.equal('<i class="fa fa-fw fa-times"></i>');
     expect(action.classList.contains('foo')).to.equal(true);
     action.querySelector('i').click();
-    expect(
-      fixture.componentInstance.events[0].actions[0].onClick
-    ).to.have.been.calledWith({ event: fixture.componentInstance.events[0] });
+    const actionSpy = fixture.componentInstance.events[0].actions[0]
+      .onClick as sinon.SinonSpy;
+    expect(actionSpy.getCall(0).args[0].event).to.equal(
+      fixture.componentInstance.events[0]
+    );
+    expect(actionSpy.getCall(0).args[0].sourceEvent instanceof MouseEvent).to.be
+      .true;
     expect(eventClicked).not.to.have.been.called;
   });
 
-  it('should allow the event width to be customised', () => {
-    const fixture: ComponentFixture<
-      CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
-    fixture.componentInstance.viewDate = new Date('2016-06-01');
-    fixture.componentInstance.events = [
-      {
-        start: new Date('2016-06-01'),
-        title: 'foo',
-        color: {
-          primary: 'blue',
-          secondary: ''
-        }
-      }
-    ];
-    fixture.componentInstance.eventWidth = 300;
-    fixture.componentInstance.ngOnChanges({
-      viewDate: {},
-      events: {},
-      eventWidth: {}
-    });
-    fixture.detectChanges();
-    expect(
-      fixture.nativeElement.querySelector('.cal-event-container').style.width
-    ).to.equal('299px');
-    fixture.destroy();
-  });
-
   it('should add a custom CSS class to days via the beforeViewRender output', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.beforeViewRender
       .pipe(take(1))
-      .subscribe(({ body }) => {
-        body.hourGrid.forEach(hour => {
-          hour.segments.forEach(segment => {
-            segment.cssClass = 'foo';
+      .subscribe(({ hourColumns }) => {
+        hourColumns.forEach(hourColumn => {
+          hourColumn.hours.forEach(hour => {
+            hour.segments.forEach(segment => {
+              segment.cssClass = 'foo';
+            });
           });
         });
       });
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
+
     fixture.detectChanges();
     expect(
       fixture.nativeElement
@@ -398,9 +328,9 @@ describe('CalendarDayViewComponent component', () => {
   });
 
   it('should resize the event by dragging from the top edge', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.events = [
       {
@@ -417,7 +347,7 @@ describe('CalendarDayViewComponent component', () => {
         }
       }
     ];
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
+
     fixture.detectChanges();
     document.body.appendChild(fixture.nativeElement);
     const event: HTMLElement = fixture.nativeElement.querySelector(
@@ -463,9 +393,9 @@ describe('CalendarDayViewComponent component', () => {
   });
 
   it('should resize the event by dragging from the bottom edge', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.events = [
       {
@@ -482,7 +412,7 @@ describe('CalendarDayViewComponent component', () => {
         }
       }
     ];
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
+
     fixture.detectChanges();
     document.body.appendChild(fixture.nativeElement);
     const event: HTMLElement = fixture.nativeElement.querySelector(
@@ -528,9 +458,9 @@ describe('CalendarDayViewComponent component', () => {
   });
 
   it('should resize back to the original edge', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.events = [
       {
@@ -547,7 +477,7 @@ describe('CalendarDayViewComponent component', () => {
         }
       }
     ];
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
+
     fixture.detectChanges();
     document.body.appendChild(fixture.nativeElement);
     const event: HTMLElement = fixture.nativeElement.querySelector(
@@ -583,9 +513,9 @@ describe('CalendarDayViewComponent component', () => {
   });
 
   it('should resize events with no end date', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.events = [
       {
@@ -599,7 +529,7 @@ describe('CalendarDayViewComponent component', () => {
         }
       }
     ];
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
+
     fixture.detectChanges();
     document.body.appendChild(fixture.nativeElement);
     const event: HTMLElement = fixture.nativeElement.querySelector(
@@ -644,9 +574,9 @@ describe('CalendarDayViewComponent component', () => {
   });
 
   it('should resize events with no end date with a custom amount of segments', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.hourSegments = 4;
     fixture.componentInstance.events = [
@@ -661,7 +591,7 @@ describe('CalendarDayViewComponent component', () => {
         }
       }
     ];
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
+
     fixture.detectChanges();
     document.body.appendChild(fixture.nativeElement);
     const event: HTMLElement = fixture.nativeElement.querySelector(
@@ -707,9 +637,9 @@ describe('CalendarDayViewComponent component', () => {
   });
 
   it('should resize the event and respect the event snap size', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.events = [
       {
@@ -727,7 +657,7 @@ describe('CalendarDayViewComponent component', () => {
       }
     ];
     fixture.componentInstance.eventSnapSize = 1;
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
+
     fixture.detectChanges();
     document.body.appendChild(fixture.nativeElement);
     const event: HTMLElement = fixture.nativeElement.querySelector(
@@ -773,9 +703,9 @@ describe('CalendarDayViewComponent component', () => {
   });
 
   it('should show a tooltip on mouseover of the event', fakeAsync(() => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     eventTitle.dayTooltip = (e: CalendarEvent) => {
       return `title: ${e.title}`;
     };
@@ -791,7 +721,7 @@ describe('CalendarDayViewComponent component', () => {
         }
       }
     ];
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
+
     fixture.detectChanges();
     const event: HTMLElement = fixture.nativeElement.querySelector(
       '.cal-event'
@@ -811,12 +741,13 @@ describe('CalendarDayViewComponent component', () => {
     triggerDomEvent('mouseleave', event);
     fixture.detectChanges();
     expect(!!document.body.querySelector('.cal-tooltip')).to.equal(false);
+    fixture.destroy();
   }));
 
   it('should disable the tooltip', fakeAsync(() => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     eventTitle.dayTooltip = () => '';
     fixture.componentInstance.viewDate = new Date('2016-06-01');
     fixture.componentInstance.events = [
@@ -830,7 +761,7 @@ describe('CalendarDayViewComponent component', () => {
         }
       }
     ];
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
+
     fixture.detectChanges();
     const event: HTMLElement = fixture.nativeElement.querySelector(
       '.cal-event'
@@ -839,12 +770,13 @@ describe('CalendarDayViewComponent component', () => {
     fixture.detectChanges();
     flush();
     expect(!!document.body.querySelector('.cal-tooltip')).to.equal(false);
+    fixture.destroy();
   }));
 
   it('should allow events to be dragged and dropped', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.events = [
       {
@@ -861,7 +793,7 @@ describe('CalendarDayViewComponent component', () => {
         draggable: true
       }
     ];
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
+
     fixture.detectChanges();
     document.body.appendChild(fixture.nativeElement);
     const event: HTMLElement = fixture.nativeElement.querySelector(
@@ -906,84 +838,10 @@ describe('CalendarDayViewComponent component', () => {
     expect(eventDropped).to.have.been.calledOnce;
   });
 
-  it('should not allow events to be dragged outside of the calendar', () => {
-    const fixture: ComponentFixture<
-      CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
-    fixture.componentInstance.viewDate = new Date('2016-06-27');
-    fixture.componentInstance.events = [
-      {
-        title: 'foo',
-        color: { primary: '', secondary: '' },
-        start: moment('2016-06-27')
-          .startOf('day')
-          .add(4, 'hours')
-          .toDate(),
-        end: moment('2016-06-27')
-          .startOf('day')
-          .add(6, 'hours')
-          .toDate(),
-        draggable: true
-      }
-    ];
-    fixture.componentInstance.eventSnapSize = 1;
-    const eventDropped = sinon.spy();
-    fixture.componentInstance.eventTimesChanged.subscribe(eventDropped);
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
-    fixture.detectChanges();
-    document.body.appendChild(fixture.nativeElement);
-    const event: HTMLElement = fixture.nativeElement.querySelector(
-      '.cal-event-container'
-    );
-    const eventPosition: ClientRect = event.getBoundingClientRect();
-    const calendarPosition: ClientRect = fixture.nativeElement.getBoundingClientRect();
-    triggerDomEvent('mousedown', event, {
-      clientY: eventPosition.top,
-      clientX: eventPosition.left + 10
-    });
-    fixture.detectChanges();
-    triggerDomEvent('mousemove', document.body, {
-      clientY: calendarPosition.top,
-      clientX: eventPosition.left + 10
-    });
-    fixture.detectChanges();
-    expect(event.getBoundingClientRect().top).to.equal(calendarPosition.top);
-    expect(event.getBoundingClientRect().bottom).to.equal(
-      calendarPosition.top + eventPosition.height
-    );
-    triggerDomEvent('mousemove', document.body, {
-      clientY: calendarPosition.top - 60,
-      clientX: eventPosition.left + 10
-    });
-    fixture.detectChanges();
-    expect(event.getBoundingClientRect().top).to.equal(calendarPosition.top);
-    expect(event.getBoundingClientRect().bottom).to.equal(
-      calendarPosition.top + eventPosition.height
-    );
-    triggerDomEvent('mouseup', document.body, {
-      clientY: calendarPosition.top - 60,
-      clientX: eventPosition.left + 10
-    });
-    fixture.detectChanges();
-    const { newStart, newEnd } = eventDropped.getCall(0).args[0];
-    expect(newStart).to.deep.equal(
-      moment('2016-06-27')
-        .startOf('day')
-        .toDate()
-    );
-    expect(newEnd).to.deep.equal(
-      moment('2016-06-27')
-        .startOf('day')
-        .add(2, 'hours')
-        .toDate()
-    );
-    fixture.destroy();
-  });
-
   it('should allow events to be dragged outside of the calendar', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.events = [
       {
@@ -1003,7 +861,7 @@ describe('CalendarDayViewComponent component', () => {
     fixture.componentInstance.snapDraggedEvents = false;
     const eventDropped = sinon.spy();
     fixture.componentInstance.eventTimesChanged.subscribe(eventDropped);
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
+
     fixture.detectChanges();
     document.body.appendChild(fixture.nativeElement);
     const event: HTMLElement = fixture.nativeElement.querySelector(
@@ -1055,9 +913,9 @@ describe('CalendarDayViewComponent component', () => {
   });
 
   it('should round event drag sizes to the event snap size when dragging and dropping non snapped events', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.events = [
       {
@@ -1077,7 +935,7 @@ describe('CalendarDayViewComponent component', () => {
     fixture.componentInstance.snapDraggedEvents = false;
     const eventDropped = sinon.spy();
     fixture.componentInstance.eventTimesChanged.subscribe(eventDropped);
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
+
     fixture.detectChanges();
     document.body.appendChild(fixture.nativeElement);
     const event: HTMLElement = fixture.nativeElement.querySelector(
@@ -1116,9 +974,9 @@ describe('CalendarDayViewComponent component', () => {
   });
 
   it('should not allow events to be resized outside of the container', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.events = [
       {
@@ -1135,7 +993,7 @@ describe('CalendarDayViewComponent component', () => {
         }
       }
     ];
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
+
     fixture.detectChanges();
     document.body.appendChild(fixture.nativeElement);
     const event: HTMLElement = fixture.nativeElement.querySelector(
@@ -1172,68 +1030,10 @@ describe('CalendarDayViewComponent component', () => {
     fixture.destroy();
   });
 
-  it('should allow external events to be dropped on the day view', () => {
-    const fixture: ComponentFixture<
-      CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
-    fixture.componentInstance.viewDate = new Date('2016-06-27');
-    fixture.componentInstance.events = [];
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
-    fixture.detectChanges();
-    document.body.appendChild(fixture.nativeElement);
-
-    const externalEventFixture: ComponentFixture<
-      ExternalEventComponent
-    > = TestBed.createComponent(ExternalEventComponent);
-    externalEventFixture.detectChanges();
-    document.body.appendChild(externalEventFixture.nativeElement);
-
-    const event: HTMLElement = externalEventFixture.nativeElement.querySelector(
-      '.external-event'
-    );
-    const eventPosition: ClientRect = event.getBoundingClientRect();
-
-    const segments: any[] = Array.from(
-      fixture.nativeElement.querySelectorAll('.cal-hour-segment')
-    );
-    const segment: HTMLElement = segments[2].parentNode;
-    const segmentPosition: ClientRect = segment.getBoundingClientRect();
-
-    const eventDropped: sinon.SinonSpy = sinon.spy();
-    fixture.componentInstance.eventTimesChanged.subscribe(eventDropped);
-    triggerDomEvent('mousedown', event, {
-      clientY: eventPosition.top,
-      clientX: eventPosition.left
-    });
-    fixture.detectChanges();
-    triggerDomEvent('mousemove', document.body, {
-      clientY: segmentPosition.top,
-      clientX: segmentPosition.left
-    });
-    fixture.detectChanges();
-    expect(segment.classList.contains('cal-drag-over')).to.equal(true);
-    triggerDomEvent('mouseup', document.body, {
-      clientY: segmentPosition.top,
-      clientX: segmentPosition.left
-    });
-    fixture.detectChanges();
-    fixture.destroy();
-    externalEventFixture.destroy();
-    expect(eventDropped).to.have.been.calledWith({
-      type: 'drop',
-      event: externalEventFixture.componentInstance.event,
-      newStart: moment('2016-06-27')
-        .startOf('day')
-        .add(1, 'hours')
-        .toDate(),
-      allDay: false
-    });
-  });
-
   it('should respect the event snap size when dragging and dropping', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.events = [
       {
@@ -1251,7 +1051,7 @@ describe('CalendarDayViewComponent component', () => {
       }
     ];
     fixture.componentInstance.eventSnapSize = 1;
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
+
     fixture.detectChanges();
     document.body.appendChild(fixture.nativeElement);
     const event: HTMLElement = fixture.nativeElement.querySelector(
@@ -1299,28 +1099,27 @@ describe('CalendarDayViewComponent component', () => {
 
   it('should log on invalid events', () => {
     const stub = sinon.stub(console, 'warn');
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     fixture.componentInstance.events = [
-      { start: '2017-01-01', title: '', color: { primary: '', secondary: '' } }
+      { start: 1234, title: '', color: { primary: '', secondary: '' } }
     ] as any;
-    fixture.componentInstance.ngOnChanges({ events: {} });
+    fixture.componentInstance.viewDate = new Date('2017-01-01');
     fixture.detectChanges();
     stub.restore();
     expect(stub).to.have.been.calledOnce; // tslint:disable-line
   });
 
   it('should allow the hour segment height to be customised', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     fixture.componentInstance.hourSegmentHeight = 45;
     fixture.componentInstance.viewDate = new Date('2016-06-01');
-    fixture.componentInstance.ngOnChanges({ viewDate: {} });
     fixture.detectChanges();
     expect(
-      fixture.nativeElement.querySelector('mwl-calendar-day-view-hour-segment')
+      fixture.nativeElement.querySelector('mwl-calendar-week-view-hour-segment')
         .style.height
     ).to.equal('45px');
     expect(
@@ -1329,12 +1128,12 @@ describe('CalendarDayViewComponent component', () => {
   });
 
   it('should only call the beforeViewRender output once when refreshing the view', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     fixture.componentInstance.refresh = new Subject();
-    fixture.componentInstance.ngOnInit();
     fixture.componentInstance.viewDate = new Date('2016-06-27');
+    fixture.detectChanges();
     const beforeViewRenderCalled = sinon.spy();
     // use subscription to test that it was only called a max of one times
     const subscription = fixture.componentInstance.beforeViewRender.subscribe(
@@ -1347,35 +1146,35 @@ describe('CalendarDayViewComponent component', () => {
   });
 
   it('should only call the beforeViewRender output once when changing the view date', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
-    fixture.componentInstance.ngOnInit();
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
-    fixture.componentInstance.ngOnChanges({ viewDate: {} });
+    fixture.detectChanges();
     const beforeViewRenderCalled = sinon.spy();
     // use subscription to test that it was only called a max of one times
     const subscription = fixture.componentInstance.beforeViewRender.subscribe(
       beforeViewRenderCalled
     );
     fixture.componentInstance.viewDate = new Date('2016-06-28');
-    fixture.componentInstance.ngOnChanges({ viewDate: {} });
+    fixture.detectChanges();
     expect(beforeViewRenderCalled).to.have.been.calledOnce;
     subscription.unsubscribe();
     fixture.destroy();
   });
 
   it('should expose the view period on the beforeViewRender output', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     const beforeViewRenderCalled = sinon.spy();
     fixture.componentInstance.beforeViewRender
       .pipe(take(1))
       .subscribe(beforeViewRenderCalled);
-    fixture.componentInstance.ngOnInit();
+
     fixture.componentInstance.viewDate = new Date('2016-06-27');
-    fixture.componentInstance.ngOnChanges({ viewDate: {} });
+    fixture.detectChanges();
+
     const { period } = beforeViewRenderCalled.getCall(0).args[0];
     expect(period.start).to.be.an.instanceOf(Date);
     expect(period.end).to.be.an.instanceOf(Date);
@@ -1384,14 +1183,14 @@ describe('CalendarDayViewComponent component', () => {
   });
 
   it('should expose the events on the beforeViewRender output', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     const beforeViewRenderCalled = sinon.spy();
     fixture.componentInstance.beforeViewRender
       .pipe(take(1))
       .subscribe(beforeViewRenderCalled);
-    fixture.componentInstance.ngOnInit();
+
     fixture.componentInstance.events = [
       {
         start: new Date('2016-05-30'),
@@ -1403,22 +1202,30 @@ describe('CalendarDayViewComponent component', () => {
         end: new Date('2016-06-02'),
         title: 'bar',
         allDay: true
+      },
+      {
+        start: new Date('2017-05-30'),
+        end: new Date('2017-06-02'),
+        title: 'bar',
+        allDay: true
       }
     ];
     fixture.componentInstance.viewDate = new Date('2016-06-01');
-    fixture.componentInstance.ngOnChanges({ viewDate: {} });
+    fixture.detectChanges();
+
     const {
-      period: { events },
-      body: { allDayEvents }
+      period: { events }
     } = beforeViewRenderCalled.getCall(0).args[0];
-    expect(events).to.deep.equal([fixture.componentInstance.events[0]]);
-    expect(allDayEvents).to.deep.equal([fixture.componentInstance.events[1]]);
+    expect(events).to.deep.equal([
+      fixture.componentInstance.events[0],
+      fixture.componentInstance.events[1]
+    ]);
   });
 
   it('should drag an all day event onto the time grid', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2018-07-29');
     fixture.componentInstance.events = [
       {
@@ -1436,12 +1243,10 @@ describe('CalendarDayViewComponent component', () => {
       }
     ];
     fixture.componentInstance.snapDraggedEvents = false;
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
+
     fixture.detectChanges();
     document.body.appendChild(fixture.nativeElement);
-    const event = fixture.nativeElement.querySelector(
-      '.cal-all-day-events mwl-calendar-day-view-event'
-    );
+    const event = fixture.nativeElement.querySelector('.cal-event-container');
     const rect: ClientRect = event.getBoundingClientRect();
     let dragEvent: CalendarEventTimesChangedEvent;
     fixture.componentInstance.eventTimesChanged.subscribe(e => {
@@ -1453,7 +1258,7 @@ describe('CalendarDayViewComponent component', () => {
     });
     fixture.detectChanges();
     const hourSegment = fixture.nativeElement.querySelectorAll(
-      'mwl-calendar-day-view-hour-segment'
+      '.cal-day-columns mwl-calendar-week-view-hour-segment'
     )[3];
     const hourSegmentPosition = hourSegment.getBoundingClientRect();
     triggerDomEvent('mousemove', hourSegment, {
@@ -1478,71 +1283,6 @@ describe('CalendarDayViewComponent component', () => {
     });
   });
 
-  it('should drag a time event onto the all day events', () => {
-    const fixture: ComponentFixture<
-      CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
-    fixture.componentInstance.viewDate = new Date('2018-07-29');
-    fixture.componentInstance.events = [
-      {
-        start: moment(new Date('2018-07-29'))
-          .startOf('day')
-          .add(3, 'hours')
-          .toDate(),
-        end: moment(new Date('2018-07-29'))
-          .startOf('day')
-          .add(18, 'hours')
-          .toDate(),
-        title: 'foo',
-        draggable: true,
-        allDay: false
-      },
-      {
-        start: new Date('2018-07-29'),
-        allDay: true,
-        title: 'bar'
-      }
-    ];
-    fixture.componentInstance.snapDraggedEvents = false;
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
-    fixture.detectChanges();
-    document.body.appendChild(fixture.nativeElement);
-    const event = fixture.nativeElement.querySelector('.cal-event-container');
-    const rect: ClientRect = event.getBoundingClientRect();
-    let dragEvent: CalendarEventTimesChangedEvent;
-    fixture.componentInstance.eventTimesChanged.subscribe(e => {
-      dragEvent = e;
-    });
-    triggerDomEvent('mousedown', event, {
-      clientX: rect.left,
-      clientY: rect.top
-    });
-    fixture.detectChanges();
-    const allDayEvents = fixture.nativeElement.querySelector(
-      '.cal-all-day-events'
-    );
-    const allDayEventsPosition = allDayEvents.getBoundingClientRect();
-    triggerDomEvent('mousemove', allDayEvents, {
-      clientX: allDayEventsPosition.left,
-      clientY: allDayEventsPosition.top
-    });
-    fixture.detectChanges();
-    triggerDomEvent('mouseup', allDayEvents, {
-      clientX: allDayEventsPosition.left,
-      clientY: allDayEventsPosition.top
-    });
-    fixture.detectChanges();
-    fixture.destroy();
-    expect(dragEvent).to.deep.equal({
-      type: 'drop',
-      allDay: true,
-      event: fixture.componentInstance.events[0],
-      newStart: moment(new Date('2018-07-29'))
-        .startOf('day')
-        .toDate()
-    });
-  });
-
   it('should allow css variables for colors', () => {
     const style = document.createElement('style');
     style.setAttribute('type', 'text/css');
@@ -1553,10 +1293,10 @@ describe('CalendarDayViewComponent component', () => {
       }
     `;
     document.head.appendChild(style);
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(
       CalendarDayViewComponent
-    > = TestBed.createComponent(CalendarDayViewComponent);
-    fixture.componentInstance.ngOnInit();
+    );
+
     fixture.componentInstance.viewDate = new Date('2016-06-01');
     fixture.componentInstance.events = [
       {
@@ -1569,7 +1309,7 @@ describe('CalendarDayViewComponent component', () => {
         }
       }
     ];
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
+
     fixture.detectChanges();
 
     const computedStyles: CSSStyleDeclaration = window.getComputedStyle(
