@@ -29,6 +29,10 @@ import { Subject } from 'rxjs';
 import { triggerDomEvent } from './util';
 import { take } from 'rxjs/operators';
 import { adapterFactory } from '../src/date-adapters/date-fns';
+import localeDe from '@angular/common/locales/de';
+import { registerLocaleData } from '@angular/common';
+
+registerLocaleData(localeDe);
 
 describe('calendarMonthView component', () => {
   beforeEach(() => {
@@ -58,9 +62,9 @@ describe('calendarMonthView component', () => {
   }));
 
   it('should generate the month view', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.ngOnChanges({ viewDate: {} });
     expect(fixture.componentInstance.view.rowOffsets).to.deep.equal([
@@ -79,14 +83,15 @@ describe('calendarMonthView component', () => {
   });
 
   it('should emit on the columnHeaderClicked output', done => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-29');
     fixture.componentInstance.ngOnChanges({ viewDate: {} });
     fixture.detectChanges();
     fixture.componentInstance.columnHeaderClicked.subscribe(val => {
-      expect(val).to.equal(0);
+      expect(val.isoDayNumber).to.equal(0);
+      expect(val.sourceEvent instanceof MouseEvent).to.be.true;
       done();
     });
     fixture.nativeElement.querySelector('.cal-header .cal-cell').click();
@@ -94,9 +99,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should generate the week view with default colors for events', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.ngOnInit();
     fixture.componentInstance.viewDate = new Date('2016-06-01');
     fixture.componentInstance.events = [
@@ -125,9 +130,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should generate the month view without from week excluded days', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-01-10');
     fixture.componentInstance.excludeDays = [0, 6];
     fixture.componentInstance.ngOnChanges({ viewDate: {} });
@@ -147,9 +152,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should update the month view when excluded days changed', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-01-10');
     fixture.componentInstance.excludeDays = [0, 1, 2];
     fixture.componentInstance.ngOnChanges({ excludeDays: {} });
@@ -159,9 +164,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should open and close the active day events list', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     expect(fixture.componentInstance.openRowIndex).to.equal(undefined);
     expect(fixture.componentInstance.openDay).to.equal(undefined);
     fixture.componentInstance.viewDate = moment()
@@ -189,9 +194,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should use the activeDay input instead of the viewDate to determine the active day', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     expect(fixture.componentInstance.openRowIndex).to.equal(undefined);
     expect(fixture.componentInstance.openDay).to.equal(undefined);
     fixture.componentInstance.viewDate = moment()
@@ -224,9 +229,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should add a custom CSS class to events', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.events = [
       {
@@ -250,9 +255,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should add a custom CSS class to days via the beforeViewRender output', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.beforeViewRender
       .pipe(take(1))
@@ -270,9 +275,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should add a custom CSS class to headers via the beforeViewRender output', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.beforeViewRender
       .pipe(take(1))
@@ -290,9 +295,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should not remove other classes when removing the cssClass', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     let firstDay: CalendarMonthViewDay;
     fixture.componentInstance.beforeViewRender
@@ -316,9 +321,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should add the highlight class to events on mouse over', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.events = [
       {
@@ -350,9 +355,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should add event actions to the active day events', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.events = [
       {
@@ -381,15 +386,59 @@ describe('calendarMonthView component', () => {
     expect(action.innerHTML).to.equal('<i class="fa fa-fw fa-times"></i>');
     expect(action.classList.contains('foo')).to.equal(true);
     action.click();
+    const actionSpy = fixture.componentInstance.events[0].actions[0]
+      .onClick as sinon.SinonSpy;
+    expect(actionSpy.getCall(0).args[0].event).to.equal(
+      fixture.componentInstance.events[0]
+    );
+    expect(actionSpy.getCall(0).args[0].sourceEvent instanceof MouseEvent).to.be
+      .true;
+  });
+
+  it('should add event actions to the active day events on enter keypress', () => {
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
+      CalendarMonthViewComponent
+    );
+    fixture.componentInstance.viewDate = new Date('2016-06-27');
+    fixture.componentInstance.events = [
+      {
+        start: new Date('2016-06-26'),
+        end: new Date('2016-06-28'),
+        title: 'foo',
+        color: {
+          primary: 'blue',
+          secondary: 'rgb(238, 238, 238)'
+        },
+        actions: [
+          {
+            label: '<i class="fa fa-fw fa-times"></i>',
+            onClick: spy(),
+            cssClass: 'foo'
+          }
+        ]
+      }
+    ];
+    fixture.componentInstance.activeDayIsOpen = true;
+    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
+    fixture.detectChanges();
+    const action: HTMLElement = fixture.nativeElement.querySelector(
+      '.cal-open-day-events .cal-event-action'
+    );
+    expect(action.innerHTML).to.equal('<i class="fa fa-fw fa-times"></i>');
+    expect(action.classList.contains('foo')).to.equal(true);
+    const sourceEvent = triggerDomEvent('keydown', action, { keyCode: 13 });
     expect(
       fixture.componentInstance.events[0].actions[0].onClick
-    ).to.have.been.calledWith({ event: fixture.componentInstance.events[0] });
+    ).to.have.been.calledWith({
+      event: fixture.componentInstance.events[0],
+      sourceEvent
+    });
   });
 
   it('should call the event clicked callback', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.events = [
       {
@@ -410,15 +459,50 @@ describe('calendarMonthView component', () => {
     );
     expect(title.innerHTML).to.equal('<span>foo</span>');
     fixture.componentInstance.eventClicked.subscribe(val => {
-      expect(val).to.deep.equal({ event: fixture.componentInstance.events[0] });
+      expect(val).to.deep.equal({
+        event: fixture.componentInstance.events[0],
+        sourceEvent: window['event']
+      });
     });
     title.click();
   });
 
-  it('should refresh the view when the refresh observable is emitted on', () => {
-    const fixture: ComponentFixture<
+  it('should call the event clicked callback on enter keypress', () => {
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
+    fixture.componentInstance.viewDate = new Date('2016-06-27');
+    fixture.componentInstance.events = [
+      {
+        start: new Date('2016-06-26'),
+        end: new Date('2016-06-28'),
+        title: '<span>foo</span>',
+        color: {
+          primary: 'blue',
+          secondary: 'rgb(238, 238, 238)'
+        }
+      }
+    ];
+    fixture.componentInstance.activeDayIsOpen = true;
+    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
+    fixture.detectChanges();
+    const title: HTMLElement = fixture.nativeElement.querySelector(
+      '.cal-open-day-events .cal-event-title'
+    );
+    expect(title.innerHTML).to.equal('<span>foo</span>');
+    fixture.componentInstance.eventClicked.subscribe(val => {
+      expect(val).to.deep.equal({
+        event: fixture.componentInstance.events[0],
+        sourceEvent: window['event']
+      });
+    });
+    triggerDomEvent('keydown', title, { keyCode: 13 });
+  });
+
+  it('should refresh the view when the refresh observable is emitted on', () => {
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
+      CalendarMonthViewComponent
+    );
     fixture.componentInstance.refresh = new Subject();
     fixture.componentInstance.ngOnInit();
     fixture.componentInstance.viewDate = new Date('2016-06-27');
@@ -441,9 +525,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should allow the event title to be customised by the calendarConfig provider', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     eventTitle.month = (event: CalendarEvent) => {
       return `foo ${event.title}`;
     };
@@ -469,9 +553,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should allow the locale to be changed', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.locale = 'de';
     fixture.componentInstance.viewDate = new Date();
     fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
@@ -484,9 +568,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should allow the badge total to be customised', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.beforeViewRender
       .pipe(take(1))
@@ -502,9 +586,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should show a tooltip on mouseover of the event', fakeAsync(() => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     eventTitle.monthTooltip = (e: CalendarEvent) => {
       return `title: ${e.title}`;
     };
@@ -543,9 +627,9 @@ describe('calendarMonthView component', () => {
   }));
 
   it('should show a tooltip on mouseover of the event after a delay', fakeAsync(() => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     eventTitle.monthTooltip = (e: CalendarEvent) => {
       return `title: ${e.title}`;
     };
@@ -584,9 +668,9 @@ describe('calendarMonthView component', () => {
   }));
 
   it('should disable the tooltip', fakeAsync(() => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     eventTitle.monthTooltip = () => '';
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.events = [
@@ -612,9 +696,9 @@ describe('calendarMonthView component', () => {
   }));
 
   it('should allow the start of the week to be changed', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.weekStartsOn = 1;
     fixture.componentInstance.ngOnChanges({ viewDate: {} });
@@ -626,9 +710,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should allow events to be dragged and dropped', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-12-05');
     fixture.componentInstance.events = [
       {
@@ -695,9 +779,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should apply the year, month and date changes in the correct order when dragging and dropping events', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2017-02-05');
     fixture.componentInstance.events = [
       {
@@ -746,9 +830,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should not fire drop events when dropping on the source date', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2018-10-13');
     fixture.componentInstance.events = [
       {
@@ -792,9 +876,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should update the event title', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-01');
     fixture.componentInstance.events = [
       {
@@ -821,9 +905,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should handle the click event on month cell events', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.events = [
       {
@@ -850,17 +934,17 @@ describe('calendarMonthView component', () => {
       '.cal-days .cal-cell-row .cal-cell:nth-child(4) .cal-events .cal-event'
     );
     event.click();
-    fixture.destroy();
-    expect(eventClickedEvent).to.deep.equal({
-      event: fixture.componentInstance.events[0]
-    });
+    expect(eventClickedEvent.event).to.equal(
+      fixture.componentInstance.events[0]
+    );
+    expect(eventClickedEvent.sourceEvent instanceof MouseEvent).to.be.true;
     expect(dayClickedFired).to.equal(false);
   });
 
   it('should add helper classes to the header cells', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.ngOnChanges({ viewDate: {} });
     fixture.detectChanges();
@@ -877,9 +961,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should allow the weekend days to be customised', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2017-06-25');
     fixture.componentInstance.weekendDays = [
       DAYS_OF_WEEK.FRIDAY,
@@ -900,9 +984,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should only call the beforeViewRender output once when refreshing the view', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.refresh = new Subject();
     fixture.componentInstance.ngOnInit();
     fixture.componentInstance.viewDate = new Date('2016-06-27');
@@ -919,9 +1003,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should only call the beforeViewRender output once when changing the view date', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.ngOnInit();
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.ngOnChanges({ viewDate: {} });
@@ -939,12 +1023,12 @@ describe('calendarMonthView component', () => {
 
   it('should log on invalid events', () => {
     const stub = sinon.stub(console, 'warn');
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2017-01-01');
     fixture.componentInstance.events = [
-      { start: '2017-01-01', title: '', color: { primary: '', secondary: '' } }
+      { start: 1234, title: '', color: { primary: '', secondary: '' } }
     ] as any;
     fixture.componentInstance.ngOnChanges({ events: {}, viewDate: {} });
     fixture.detectChanges();
@@ -953,9 +1037,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should expose the view period on the beforeViewRender output', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     const beforeViewRenderCalled = sinon.spy();
     fixture.componentInstance.beforeViewRender
       .pipe(take(1))
@@ -985,9 +1069,9 @@ describe('calendarMonthView component', () => {
       }
     `;
     document.head.appendChild(style);
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.ngOnInit();
     fixture.componentInstance.viewDate = new Date('2016-06-01');
     fixture.componentInstance.events = [
@@ -1014,9 +1098,9 @@ describe('calendarMonthView component', () => {
   });
 
   it('should allow the tooltip text to be updated dynamically', fakeAsync(() => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarMonthViewComponent> = TestBed.createComponent(
       CalendarMonthViewComponent
-    > = TestBed.createComponent(CalendarMonthViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.events = [
       {
