@@ -26,6 +26,25 @@ import * as sinon from 'sinon';
 import { triggerDomEvent, ExternalEventComponent } from './util';
 import { take } from 'rxjs/operators';
 import { adapterFactory } from '../src/date-adapters/date-fns';
+import { Component } from '@angular/core';
+import { By } from '@angular/platform-browser';
+import * as lolex from 'lolex';
+
+@Component({
+  template: `
+    <mwl-calendar-week-view
+      [viewDate]="viewDate"
+      [events]="events"
+      (eventTimesChanged)="eventTimesChanged($event)"
+    ></mwl-calendar-week-view>
+    <mwl-external-event></mwl-external-event>
+  `
+})
+class TestComponent {
+  viewDate: Date;
+  events: CalendarEvent[];
+  eventTimesChanged = sinon.spy();
+}
 
 describe('calendarWeekView component', () => {
   beforeEach(() => {
@@ -45,7 +64,7 @@ describe('calendarWeekView component', () => {
         ),
         DragAndDropModule
       ],
-      declarations: [ExternalEventComponent],
+      declarations: [ExternalEventComponent, TestComponent],
       providers: [{ provide: MOMENT, useValue: moment }]
     });
   });
@@ -56,9 +75,9 @@ describe('calendarWeekView component', () => {
   }));
 
   it('should generate the week view', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-29');
     fixture.componentInstance.ngOnChanges({ viewDate: {} });
     expect(fixture.componentInstance.days.length).to.equal(7);
@@ -68,9 +87,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should generate the week view without excluded days', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-29');
     fixture.componentInstance.excludeDays = [0, 6];
     fixture.componentInstance.ngOnChanges({ viewDate: {} });
@@ -79,9 +98,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should update the week view when excluded days changed', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-29');
     fixture.componentInstance.excludeDays = [0, 6];
     fixture.componentInstance.ngOnChanges({ excludeDays: {} });
@@ -100,9 +119,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should support excluding non consecutive days', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-29');
     fixture.componentInstance.excludeDays = [4, 6];
     fixture.componentInstance.ngOnChanges({ viewDate: {} });
@@ -111,9 +130,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should support excluding all but 1 day', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-29');
     fixture.componentInstance.excludeDays = [0, 1, 2, 3, 4, 5];
     fixture.componentInstance.ngOnChanges({ viewDate: {} });
@@ -122,9 +141,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should generate the week view with default colors for events', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.ngOnInit();
     fixture.componentInstance.viewDate = new Date('2016-06-01');
     fixture.componentInstance.events = [
@@ -152,25 +171,27 @@ describe('calendarWeekView component', () => {
     fixture.destroy();
   });
 
-  it('should emit on the dayHeaderClicked output', () => {
-    const fixture: ComponentFixture<
+  it('should emit on the dayHeaderClicked output', done => {
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-29');
     fixture.componentInstance.ngOnChanges({ viewDate: {} });
     fixture.detectChanges();
     fixture.componentInstance.dayHeaderClicked.subscribe(val => {
       expect(val).to.deep.equal({
-        day: fixture.componentInstance.days[0]
+        day: fixture.componentInstance.days[0],
+        sourceEvent: window['event']
       });
+      done();
     });
     fixture.nativeElement.querySelector('.cal-header').click();
   });
 
   it('should add a custom CSS class to events', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-01');
     fixture.componentInstance.events = [
       {
@@ -194,10 +215,10 @@ describe('calendarWeekView component', () => {
     fixture.destroy();
   });
 
-  it('should call the event clicked callback', () => {
-    const fixture: ComponentFixture<
+  it('should call the event clicked callback', done => {
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-01');
     fixture.componentInstance.events = [
       {
@@ -217,15 +238,19 @@ describe('calendarWeekView component', () => {
     );
     expect(title.innerHTML).to.equal('<span>foo</span>');
     fixture.componentInstance.eventClicked.subscribe(val => {
-      expect(val).to.deep.equal({ event: fixture.componentInstance.events[0] });
+      expect(val).to.deep.equal({
+        event: fixture.componentInstance.events[0],
+        sourceEvent: window['event']
+      });
+      done();
     });
     title.click();
   });
 
   it('should refresh the view when the refresh observable is emitted on', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.refresh = new Subject();
     fixture.componentInstance.ngOnInit();
     fixture.componentInstance.viewDate = new Date('2016-06-01');
@@ -245,9 +270,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should allow the event title to be customised', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     eventTitle.week = (event: CalendarEvent) => {
       return `foo ${event.title}`;
     };
@@ -272,9 +297,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should allow the locale to be changed', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.locale = 'de';
     fixture.componentInstance.viewDate = new Date();
     fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
@@ -285,9 +310,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should show a tooltip on mouseover of the event', fakeAsync(() => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     eventTitle.weekTooltip = (e: CalendarEvent) => {
       return `title: ${e.title}`;
     };
@@ -323,12 +348,13 @@ describe('calendarWeekView component', () => {
     triggerDomEvent('mouseleave', event);
     fixture.detectChanges();
     expect(!!document.body.querySelector('.cal-tooltip')).to.equal(false);
+    fixture.destroy();
   }));
 
   it('should disable the tooltip', fakeAsync(() => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     eventTitle.weekTooltip = () => '';
     fixture.componentInstance.viewDate = new Date('2016-06-01');
     fixture.componentInstance.events = [
@@ -351,12 +377,13 @@ describe('calendarWeekView component', () => {
     fixture.detectChanges();
     flush();
     expect(!!document.body.querySelector('.cal-tooltip')).to.equal(false);
+    fixture.destroy();
   }));
 
   it('should allow the start of the week to be changed', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.weekStartsOn = 1;
     fixture.componentInstance.ngOnChanges({ viewDate: {} });
@@ -368,9 +395,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should resize the all day event by dragging from the left edge', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.events = [
       {
@@ -437,9 +464,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should resize the all day event by dragging from the right edge', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.events = [
       {
@@ -508,9 +535,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should resize all day events with no end date', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.events = [
       {
@@ -576,9 +603,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should allow the all day event to be dragged and dropped', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-12-08');
     fixture.componentInstance.events = [
       {
@@ -651,9 +678,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should allow all day events to be dragged outside of the calendar', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-12-08');
     fixture.componentInstance.events = [
       {
@@ -716,9 +743,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should round all day event drag sizes to the event snap size when dragging and dropping non snapped events', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-12-08');
     fixture.componentInstance.events = [
       {
@@ -783,9 +810,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should not allow all day events to be resized smaller than 1 day', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.events = [
       {
@@ -839,9 +866,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should not allow all day events to be resized outside of the current view', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.events = [
       {
@@ -906,20 +933,14 @@ describe('calendarWeekView component', () => {
   });
 
   it('should allow external events to be dropped on the week view headers', () => {
-    const fixture: ComponentFixture<
-      CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    const fixture = TestBed.createComponent(TestComponent);
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.events = [];
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
     fixture.detectChanges();
     document.body.appendChild(fixture.nativeElement);
-
-    const externalEventFixture: ComponentFixture<
-      ExternalEventComponent
-    > = TestBed.createComponent(ExternalEventComponent);
-    externalEventFixture.detectChanges();
-    document.body.appendChild(externalEventFixture.nativeElement);
+    const externalEventFixture = fixture.debugElement.query(
+      By.directive(ExternalEventComponent)
+    );
 
     const event: HTMLElement = externalEventFixture.nativeElement.querySelector(
       '.external-event'
@@ -932,8 +953,7 @@ describe('calendarWeekView component', () => {
     const header: HTMLElement = headers[2];
     const headerPosition: ClientRect = header.getBoundingClientRect();
 
-    const eventDropped: sinon.SinonSpy = sinon.spy();
-    fixture.componentInstance.eventTimesChanged.subscribe(eventDropped);
+    const eventDropped = fixture.componentInstance.eventTimesChanged;
     triggerDomEvent('mousedown', event, {
       clientY: eventPosition.top,
       clientX: eventPosition.left
@@ -949,8 +969,6 @@ describe('calendarWeekView component', () => {
       clientX: headerPosition.left
     });
     fixture.detectChanges();
-    fixture.destroy();
-    externalEventFixture.destroy();
     expect(eventDropped).to.have.been.calledWith({
       type: 'drop',
       event: externalEventFixture.componentInstance.event,
@@ -960,12 +978,13 @@ describe('calendarWeekView component', () => {
         .toDate(),
       allDay: true
     });
+    expect(eventDropped).to.have.been.calledOnce;
   });
 
   it('should allow the weekend days to be customised', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2017-06-25');
     fixture.componentInstance.weekendDays = [
       DAYS_OF_WEEK.FRIDAY,
@@ -983,9 +1002,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should add a custom CSS class to headers via the beforeViewRender output', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.beforeViewRender
       .pipe(take(1))
@@ -1004,11 +1023,11 @@ describe('calendarWeekView component', () => {
 
   it('should log on invalid events', () => {
     const stub = sinon.stub(console, 'warn');
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.events = [
-      { start: '2017-01-01', title: '', color: { primary: '', secondary: '' } }
+      { start: 1234, title: '', color: { primary: '', secondary: '' } }
     ] as any;
     fixture.componentInstance.viewDate = new Date('2017-01-01');
     fixture.componentInstance.ngOnChanges({ events: {}, viewDate: {} });
@@ -1018,9 +1037,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should only call the beforeViewRender output once when refreshing the view', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.refresh = new Subject();
     fixture.componentInstance.ngOnInit();
     fixture.componentInstance.viewDate = new Date('2016-06-27');
@@ -1036,9 +1055,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should only call the beforeViewRender output once when changing the view date', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.ngOnInit();
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     fixture.componentInstance.ngOnChanges({ viewDate: {} });
@@ -1055,9 +1074,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should expose the view period on the beforeViewRender output', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     const beforeViewRenderCalled = sinon.spy();
     fixture.componentInstance.beforeViewRender
       .pipe(take(1))
@@ -1074,9 +1093,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should add event actions to each event', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2016-06-27');
     const eventClicked = sinon.spy();
     fixture.componentInstance.eventClicked.subscribe(eventClicked);
@@ -1102,16 +1121,20 @@ describe('calendarWeekView component', () => {
     expect(action.innerHTML).to.equal('<i class="fa fa-fw fa-times"></i>');
     expect(action.classList.contains('foo')).to.equal(true);
     action.querySelector('i').click();
-    expect(
-      fixture.componentInstance.events[0].actions[0].onClick
-    ).to.have.been.calledWith({ event: fixture.componentInstance.events[0] });
+    const actionSpy = fixture.componentInstance.events[0].actions[0]
+      .onClick as sinon.SinonSpy;
+    expect(actionSpy.getCall(0).args[0].event).to.equal(
+      fixture.componentInstance.events[0]
+    );
+    expect(actionSpy.getCall(0).args[0].sourceEvent instanceof MouseEvent).to.be
+      .true;
     expect(eventClicked).not.to.have.been.called;
   });
 
   it('should make a 4 day week', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2018-07-29');
     fixture.componentInstance.daysInWeek = 4;
     fixture.componentInstance.events = [
@@ -1142,9 +1165,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should resize a time event from the end to another day on the right', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2018-07-29');
     fixture.componentInstance.events = [
       {
@@ -1213,9 +1236,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should resize a time event from the end to another day on the left', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2018-07-29');
     fixture.componentInstance.events = [
       {
@@ -1284,9 +1307,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should resize a time event from the end and not allow it to end before it starts', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2018-07-29');
     fixture.componentInstance.events = [
       {
@@ -1354,9 +1377,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should resize a time event from the start to another day on the left', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2018-07-30');
     fixture.componentInstance.events = [
       {
@@ -1423,9 +1446,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should resize a time event from the start to another day on the right', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2018-07-30');
     fixture.componentInstance.events = [
       {
@@ -1494,9 +1517,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should resize a time event and not allow it to start after it ends', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2018-07-30');
     fixture.componentInstance.events = [
       {
@@ -1564,9 +1587,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should drag time events to different days and columns while snapping to a grid', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2018-07-29');
     fixture.componentInstance.events = [
       {
@@ -1654,9 +1677,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should drag time events back to their original position while snapping to a grid', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2018-07-29');
     const originalEvent = {
       start: moment(new Date('2018-07-29'))
@@ -1714,9 +1737,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should drag time events without end dates', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2018-07-29');
     fixture.componentInstance.events = [
       {
@@ -1780,9 +1803,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should drag time events to different days and columns while not snapping to a grid', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2018-07-29');
     fixture.componentInstance.events = [
       {
@@ -1847,9 +1870,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should drag time events back to their original position while not snapping to a grid', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2018-07-29');
     const originalEvent = {
       start: moment(new Date('2018-07-29'))
@@ -1901,9 +1924,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should drag an all day event onto the time grid', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2018-07-29');
     fixture.componentInstance.events = [
       {
@@ -1964,9 +1987,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should drag a time event onto the all day grid', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2018-07-29');
     fixture.componentInstance.events = [
       {
@@ -2036,9 +2059,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should handle time event objects changing when resizing', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2018-07-29');
     fixture.componentInstance.events = [
       {
@@ -2096,9 +2119,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should drag time events to different days and columns while snapping to a grid and excluding weekends', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2019-03-01');
     fixture.componentInstance.excludeDays = [0, 6];
     fixture.componentInstance.daysInWeek = 4;
@@ -2167,9 +2190,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should preserve css classes on hour segments when dragging an event', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2018-07-29');
     fixture.componentInstance.events = [
       {
@@ -2223,9 +2246,9 @@ describe('calendarWeekView component', () => {
   });
 
   it('should resize a time event to the minimum height', () => {
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.viewDate = new Date('2018-07-29');
     fixture.componentInstance.events = [
       {
@@ -2296,9 +2319,9 @@ describe('calendarWeekView component', () => {
       }
     `;
     document.head.appendChild(style);
-    const fixture: ComponentFixture<
+    const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
       CalendarWeekViewComponent
-    > = TestBed.createComponent(CalendarWeekViewComponent);
+    );
     fixture.componentInstance.ngOnInit();
     fixture.componentInstance.viewDate = new Date('2016-06-01');
     fixture.componentInstance.events = [
@@ -2325,5 +2348,140 @@ describe('calendarWeekView component', () => {
       'rgb(255, 255, 255)'
     );
     document.head.appendChild(style);
+  });
+
+  it('should allow external events to be dropped on the hour segments', () => {
+    const fixture = TestBed.createComponent(TestComponent);
+    fixture.componentInstance.viewDate = new Date('2016-06-27');
+    fixture.componentInstance.events = [];
+    fixture.detectChanges();
+    document.body.appendChild(fixture.nativeElement);
+    const externalEventFixture = fixture.debugElement.query(
+      By.directive(ExternalEventComponent)
+    );
+
+    const event: HTMLElement = externalEventFixture.nativeElement.querySelector(
+      '.external-event'
+    );
+    const eventPosition: ClientRect = event.getBoundingClientRect();
+
+    const segments: HTMLElement[] = Array.from(
+      fixture.nativeElement.querySelectorAll(
+        '.cal-day-columns .cal-hour-segment'
+      )
+    );
+    const segment = segments[1];
+    const segmentPosition: ClientRect = segment.getBoundingClientRect();
+
+    const eventDropped = fixture.componentInstance.eventTimesChanged;
+
+    triggerDomEvent('mousedown', event, {
+      clientY: eventPosition.top,
+      clientX: eventPosition.left
+    });
+    fixture.detectChanges();
+    triggerDomEvent('mousemove', document.body, {
+      clientY: segmentPosition.top,
+      clientX: segmentPosition.left
+    });
+    fixture.detectChanges();
+    triggerDomEvent('mouseup', document.body, {
+      clientY: segmentPosition.top,
+      clientX: segmentPosition.left
+    });
+    fixture.detectChanges();
+    expect(eventDropped).to.have.been.calledWith({
+      type: 'drop',
+      event: externalEventFixture.componentInstance.event,
+      newStart: moment('2016-06-27')
+        .startOf('week')
+        .add(30, 'minutes')
+        .toDate(),
+      allDay: false
+    });
+    expect(eventDropped).to.have.been.calledOnce;
+  });
+
+  describe('current time marker', () => {
+    let clock: any;
+    beforeEach(() => {
+      clock = lolex.install({
+        now: new Date('2019-09-30T11:30:25.288Z').getTime(),
+        toFake: ['Date']
+      });
+    });
+
+    afterEach(() => {
+      clock.uninstall();
+    });
+
+    it('should show a current time marker', () => {
+      const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
+        CalendarWeekViewComponent
+      );
+      fixture.componentInstance.viewDate = new Date();
+      fixture.componentInstance.ngOnChanges({
+        viewDate: {},
+        hourSegmentHeight: {}
+      });
+      fixture.detectChanges();
+      const marker = fixture.nativeElement.querySelector(
+        '.cal-day-columns .cal-day-column:nth-child(2) .cal-current-time-marker'
+      );
+      expect(marker.style.top).to.equal('690px');
+    });
+
+    it('should respect the start time', () => {
+      const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
+        CalendarWeekViewComponent
+      );
+      fixture.componentInstance.viewDate = new Date();
+      fixture.componentInstance.dayStartHour = 3;
+      fixture.componentInstance.dayStartMinute = 30;
+      fixture.componentInstance.ngOnChanges({
+        viewDate: {},
+        hourSegmentHeight: {}
+      });
+      fixture.detectChanges();
+      const marker = fixture.nativeElement.querySelector(
+        '.cal-day-columns .cal-day-column:nth-child(2) .cal-current-time-marker'
+      );
+      expect(marker.style.top).to.equal('480px');
+    });
+
+    it('should respect the end time', () => {
+      const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
+        CalendarWeekViewComponent
+      );
+      fixture.componentInstance.viewDate = new Date();
+      fixture.componentInstance.dayEndHour = 3;
+      fixture.componentInstance.ngOnChanges({
+        viewDate: {},
+        hourSegmentHeight: {}
+      });
+      fixture.detectChanges();
+      const marker = fixture.nativeElement.querySelector(
+        '.cal-day-columns .cal-day-column:nth-child(2) .cal-current-time-marker'
+      );
+      expect(marker).to.equal(null);
+    });
+
+    it('should respect the hour segment count and height', () => {
+      const fixture: ComponentFixture<CalendarWeekViewComponent> = TestBed.createComponent(
+        CalendarWeekViewComponent
+      );
+      fixture.componentInstance.viewDate = new Date();
+      fixture.componentInstance.hourSegments = 4;
+      fixture.componentInstance.hourSegmentHeight = 60;
+      fixture.componentInstance.ngOnChanges({
+        viewDate: {},
+        hourSegmentHeight: {}
+      });
+      fixture.detectChanges();
+      const marker = fixture.nativeElement.querySelector(
+        '.cal-day-columns .cal-day-column:nth-child(2) .cal-current-time-marker'
+      );
+      expect(marker.style.top).to.equal('2760px');
+    });
   });
 });
