@@ -16,14 +16,18 @@ export const sources = [
   {
     filename: 'styles.scss',
     contents: {
-      raw: require('!!raw-loader!./styles.scss').replace(
-        '../../../../angular-calendar/src/angular-calendar.scss',
-        '~angular-calendar/scss/angular-calendar.scss'
-      ),
-      highlighted: require('!!raw-loader!highlightjs-loader?lang=scss!./styles.scss').replace(
-        '../../../../angular-calendar/src/angular-calendar.scss',
-        '~angular-calendar/scss/angular-calendar.scss'
-      )
+      raw: {
+        default: require('!!raw-loader!./styles.scss').default.replace(
+          '../../../../angular-calendar/src/angular-calendar.scss',
+          '~angular-calendar/scss/angular-calendar.scss'
+        )
+      },
+      highlighted: {
+        default: require('!!raw-loader!highlightjs-loader?lang=scss!./styles.scss').default.replace(
+          '../../../../angular-calendar/src/angular-calendar.scss',
+          '~angular-calendar/scss/angular-calendar.scss'
+        )
+      }
     }
   },
   {
