@@ -3,7 +3,11 @@ import {
   ChangeDetectionStrategy,
   ViewEncapsulation
 } from '@angular/core';
-import { CalendarEvent, CalendarMonthViewDay } from 'angular-calendar';
+import {
+  CalendarEvent,
+  CalendarMonthViewDay,
+  CalendarView
+} from 'angular-calendar';
 import {
   subMonths,
   addMonths,
@@ -62,7 +66,7 @@ function endOfPeriod(period: CalendarPeriod, date: Date): Date {
   encapsulation: ViewEncapsulation.None
 })
 export class DemoComponent {
-  view: CalendarPeriod = 'month';
+  view: CalendarView | CalendarPeriod = CalendarView.Month;
 
   viewDate: Date = new Date();
 
