@@ -8,7 +8,8 @@ import {
   CalendarViewPeriod,
   CalendarMonthViewBeforeRenderEvent,
   CalendarWeekViewBeforeRenderEvent,
-  CalendarDayViewBeforeRenderEvent
+  CalendarDayViewBeforeRenderEvent,
+  CalendarView
 } from 'angular-calendar';
 import { colors } from '../demo-utils/colors';
 
@@ -18,7 +19,7 @@ import { colors } from '../demo-utils/colors';
   templateUrl: 'template.html'
 })
 export class DemoComponent {
-  view: string = 'month';
+  view: CalendarView = CalendarView.Month;
 
   viewDate: Date = new Date();
 
